### 2023-09-29 #609 @maRce10
baRulho
Submitting Author Name: Marcelo Araya-Salas
Submitting Author Github Handle: @maRce10
Repository:  https://github.com/maRce10/baRulho/
Version submitted: 2.1.0
Submission type: Standard

Version accepted: TBD

---

```
Package: baRulho
Title: Quantifying (Animal) Sound Degradation
Version: 2.1.0
Author: Marcelo Araya-Salas [aut, cre]
Maintainer: Marcelo Araya-Salas <marcelo.araya@ucr.ac.cr>
Description: Intended to facilitate acoustic analysis of (animal) sound transmission experiments, which typically aim to quantify changes in signal structure when transmitted in a given habitat by broadcasting and re-recording animal sounds at increasing distances. The package offers a workflow with functions to prepare the data set for analysis as well as to calculate and visualize several degradation metrics, including blur ratio, signal-to-noise ratio, excess attenuation and envelope correlation among others (Dabelsteen et al 1993 <doi:10.1121/1.406682>).
Imports: utils, stats, seewave, tuneR, fftw, methods, viridis, Sim.DiffProc, png, checkmate, cli
Depends: R (>= 3.2.1), warbleR (>= 1.1.29), ohun (>= 1.0.0)
    kableExtra,
    covr,
    formatR,
    Rraven
Repository: CRANs
Authors@R: person("Marcelo", "Araya-Salas", role = c("aut", "cre"), email = "marcelo.araya@ucr.ac.cr")
```

It estimates degradation measures from sounds that have been re-recorded in standard playback experiments. It also helps to prepare acoustic data for playback experiments.

Scientific community working with bioacoustic data. Useful to automate analysis of animal acoustic signals for a variety of research questions (e.g. behavior, ecology, evolution, conservation)

- To the best of my knowledge the only other similar tools is the [warbleR](https://marce10.github.io/warbleR/) package, which has similar functions for measuring features from annotation tables but not specifically related to sound degradation  (I am also the maintainer) except for [sig2noise](https://marce10.github.io/warbleR/reference/sig2noise.html).

Somehow continuous integration is not detected when running `pkgcheck::pkgcheck()`, but it is fully implemented [https://app.codecov.io/gh/maRce10/baRulho?branch=master](https://app.codecov.io/gh/maRce10/baRulho?branch=master)
