### 2023-09-29 #609 @noamross
Thank you for this submission @maRce10! Two quick notes before I start the process:
 - Some things in our checks are being triggered by some extraneous files committed in your repository: `.Rproj.user`, `.Rhistory`, etc. Some of these are in .gitignore, but probably were committed before that was created.  You can get git of these with the [`git rm --cached <file>`](https://stackoverflow.com/a/1274447/1757441).
 - @mpadge, it appears our package checks are not picking up the continuous integration tests in this package, which are set up using {tic}. Is this a bug of ours or due to something in the package setup?
