+++ R/align_test_files.R
@@ -16,7 +16,6 @@
 #' @return An object of the same class than 'X' with the aligned sounds from test (re-recorded) sound files.
 #' @export
 #' @name align_test_files
-#' @family prepare acoustic data
 #' @details The function aligns sounds found in re-recorded sound files (referenced in 'Y') according to a master sound file (referenced in 'X'). The function outputs an 'extended selection table' by default.
 #' @family test sound alignment
 #' @seealso \code{\link{manual_realign}}; \code{\link{find_markers}}; \code{\link{plot_aligned_sounds}}
