+++ pkgdown/_pkgdown.yml
@@ -1,3 +1,41 @@
-url: https://marce10.github.io/baRulho
 template:
   bootstrap: 5
+  
+destination: reference
+
+# website will be referenced on https://pkgs.rstudio.com/
+# Open a PR in https://github.com/rstudio/pkgs.rstudio.com
+url: https://marce10.github.io/baRulho
+
+template:
+  package: tidytemplate
+  bootstrap: 5
+  bslib:
+    primary: "#096B72"
+    navbar-background: "#e6f3fc"
+  trailing_slash_redirect: true
+   twitter:
+      creator: "@araya_salas"
+      card: summary
+      
+reference:
+- title: Prepare acoustic data
+  desc: >
+    Functions that format acoustic data for playback experiments
+  contents:
+  - align_test_files
+  - master_sound_files
+
+- title: Aligning test sound files
+  desc: >
+    Aligning test sounds to determine their time position
+  contents:
+  - align_test_files
+  - master_sound_files
+
+- title: Quantify degradation
+  desc: >
+    Functions that format acoustic data for playback experiments
+  contents:
+  - align_test_files
+  - master_sound_files
