### 2023-09-29 #609 @mpadge
@noamross @maRce10 The failing "no continuous integration" check is because it actually just checks for an `R CMD check` badge in the readme, which this package does not have. It has a lot of badges, but not that critical one. Obviously not a big issue here, and we can clearly accept regardless, but @maRce10 if you want that test to pass, just put a badge linking to your r cmd check status on the README.
