+++ README.Rmd
@@ -13,11 +13,11 @@ editor_options:
 [![Codecov test coverage](https://codecov.io/gh/maRce10/baRulho/branch/master/graph/badge.svg)](https://app.codecov.io/gh/maRce10/baRulho?branch=master)
 <!-- badges: end -->
 
-<img src="vignettes/baRulho_sticker.png" alt="baRulho logo" align="right" width = "25%" height="25%"/>
+<img src="reference/figures/baRulho_sticker.png" alt="baRulho logo" align="right" width = "25%" height="25%"/>
 
 [baRulho](https://cran.r-project.org/package=baRulho) is intended to facilitate acoustic analysis of (animal) sound transmission experiments, which typically aim to quantify changes in signal structure when transmitted in a given habitat by broadcasting and re-recording animal sounds at increasing distances. A common sequence of steps to experimentally test hypotheses related to sound transmission is depicted in the following diagram:
 
- <center><img src="vignettes/analysis_workflow.jpg" alt="analysis workflow" width="80%"></center>
+ <center><img src="reference/figures/analysis_workflow.jpg" alt="analysis workflow" width="80%"></center>
 
 &nbsp;
 
