+++ man/synth_sounds.Rd
@@ -80,7 +80,6 @@ Araya-Salas, M., & Smith-Vidaurre, G. (2017). warbleR: An R package to streamlin
 \code{\link[warbleR]{simulate_songs}} from the package warbleR.
 
 Other prepare acoustic data: 
-\code{\link{align_test_files}()},
 \code{\link{master_sound_file}()}
 }
 \author{
