### 2023-09-29 243c881
add R-CMD-check
