+++ R/internal_functions.R
@@ -1090,18 +1090,18 @@ env_FUN <- function(X, y, env.smooth, ovlp, wl, path, n.samples = 100) {
                       ssmooth = env.smooth)
   
   # thin
-  if (thinning < 1) {
-    if (length(nv) * thinning < 30) stop2("thinning is too high, no enough samples left for at least 1 sound file")
-    
-    # reduce size of envelope
-    nv2 <-
-      stats::approx(
-        x = seq(0, duration(clp), length.out = length(nv)),
-        y = nv,
-        n = round(length(nv) * thinning),
-        method = "linear"
-      )$y
-    }
+  # if (thinning < 1) {
+  #   if (length(nv) * thinning < 30) stop2("thinning is too high, no enough samples left for at least 1 sound file")
+  #   
+  #   # reduce size of envelope
+  #   nv2 <-
+  #     stats::approx(
+  #       x = seq(0, duration(clp), length.out = length(nv)),
+  #       y = nv,
+  #       n = round(length(nv) * thinning),
+  #       method = "linear"
+  #     )$y
+  #   }
   return(nv)
 }
 
