### 2023-09-29 3e33d15
add thining envelopes and spectra
