### 2023-09-29 021739f
add n.samples to plot_blur_ratio
