### 2023-09-29 #609 @maRce10
> @noamross @maRce10 The failing "no continuous integration" check is because it actually just checks for an `R CMD check` badge in the readme, which this package does not have. It has a lot of badges, but not that critical one. Obviously not a big issue here, and we can clearly accept regardless, but @maRce10 if you want that test to pass, just put a badge linking to your r cmd check status on the README.

Thanks! I added the `R CMD check` badge.

> Hey @maRce10. Long time no see (ha!)
>
> I will get started on this early next week.
>
> Just wanted to let you know that the impending US Gov shutdown may impact my ability to work on this. I am good to work all next week, but if we do shutdown and it is not resolved, I will have to stop all work activities (my serving as an editor counts as one) after the 7th and I can't resume until the shutdown has been lifted. Hopefully this all for naught and I am able to keep on working on this. But I will do what I can to make progress next week.

Thanks for letting me know!
