### 2023-10-02 3322fa1
optimizing excess attenuation
