+++ R/signal_to_noise_ratio.R
@@ -138,6 +138,10 @@ signal_to_noise_ratio <-
       )
     }
     
+    # 'mar' is needed when not using equal duration
+    if (eq.dur & is.null(mar))
+        stop2("'mar' must be supplied when 'eq.dur = FALSE'")
+      
     # set clusters for windows OS
     if (Sys.info()[1] == "Windows" & cores > 1) {
       cl <-
@@ -146,152 +150,144 @@ signal_to_noise_ratio <-
       cl <- cores
     }
     
-    # calculate all envelops with a apply function
-    envs <-
+    # calculate all RMS of envelopes with a apply function
+    rms_list <-
       warbleR:::pblapply_wrblr_int(
         X = seq_len(nrow(X)),
         pbar = pb,
         cl = cl,
         FUN = function(y) {
-          if (noise.ref == "custom") {
-            # read sound clip
-            signal <-
-              warbleR::read_sound_file(X = X,
-                                       index = y,
-                                       path = path)
-            
-            # add band-pass frequency filter
-            if (!is.null(bp)) {
-              # filter to bottom and top freq range
-              if (bp[1] == "freq.range") {
-                bp <- c(X$bottom.freq[y], X$top.freq[y])
-              }
-              
-              signal <-
-                seewave::ffilter(
-                  signal,
-                  f = sampling_rate,
-                  from = bp[1] * 1000,
-                  ovlp = 0,
-                  to = bp[2] * 1000,
-                  bandpass = TRUE,
-                  wl = wl,
-                  output = "Wave"
-                )
-            }
-            
-            # get RMS for signal
-            sig.env <-
-              seewave::env(signal,
-                           f = sampling_rate,
-                           envt = "abs",
-                           plot = FALSE)
-            
-            bg.env <- NA
-          }
           
-          if (noise.ref == "adjacent") {
-            # set margin to half of signal duration
-            if (eq.dur) {
-              mar <-
-                (X$end[y] - X$start[y])
-            } else if (is.null(mar)) {
-              stop2("'mar' must be provided when 'eq.dur = FALSE'")
-            }
+          # only calculate for non-markers and for ambient only if custom noise.ref
+          if (!X$sound.id[y] %in% c("marker", if (noise.ref != "custom") "ambient")){
             
-            # Read sound files to get sample rate and length
-            r <-
-              warbleR::read_sound_file(
-                X = X,
-                index = y,
-                from = X$start[y] - mar,
-                to = X$end[y],
-                header = TRUE,
-                path = path
-              )
-            
-            # reset time coordinates of sounds if lower than 0 o higher than duration
-            stn <- X$start[y] - mar
-            enn <- X$end[y] + mar
-            mar1 <- mar
-            
-            if (stn < 0) {
-              mar1 <- mar1 + stn
-              stn <- 0
-            }
-            
-            mar2 <- mar1 + X$end[y] - X$start[y]
-            
-            if (enn > r$samples / sampling_rate) {
-              enn <- r$samples / sampling_rate
-            }
-            
-            # read sound and margin
-            noise_sig <-
-              warbleR::read_sound_file(
-                X = X,
-                index = y,
-                from = stn,
-                to = enn,
-                path = path
-              )
-            
-            # add band-pass frequency filter
-            if (!is.null(bp)) {
-              # filter to bottom and top freq range
-              if (bp[1] == "freq.range") {
-                bp <- c(X$bottom.freq[y], X$top.freq[y])
+            if (noise.ref == "custom") {
+              # read sound clip
+              signal <-
+                warbleR::read_sound_file(X = X,
+                                         index = y,
+                                         path = path)
+              
+              # add band-pass frequency filter
+              if (!is.null(bp)) {
+                # filter to bottom and top freq range
[…]
