### 2023-10-02 fa37e91
make tests more efficient/clean
