+++ tests/testthat/test-spectrum_blur_ratio.R
@@ -4,7 +4,7 @@ test_that("using extended table and method 1", {
   X <-
     test_sounds_est[test_sounds_est$sound.files != "master.wav",]
   
-  X <- set_reference_sounds(X)
+  X <- set_reference_sounds(X, pb = FALSE)
   
   br <- spectrum_blur_ratio(X = X)
   
@@ -52,7 +52,7 @@ test_that("using extended table and return spectra", {
   X <-
     test_sounds_est[test_sounds_est$sound.files != "master.wav",]
   
-  X <- set_reference_sounds(X)
+  X <- set_reference_sounds(X, pb = FALSE)
   
   br <- spectrum_blur_ratio(X = X, spectra = TRUE)
   
