+++ tests/testthat/test-excess_attenuation.R
@@ -4,7 +4,7 @@ test_that("using extended table", {
   X <-
     test_sounds_est[test_sounds_est$sound.files != "master.wav",]
   
-  X <- set_reference_sounds(X)
+  X <- set_reference_sounds(X, pb = FALSE)
   
   ea <- excess_attenuation(X = X)
   
