+++ R/synth_sounds.R
@@ -5,7 +5,7 @@
 #' fm = FALSE, am = FALSE, am.amps = rep(c(1:4, 3:2), length.out = 11),
 #' mar = 0.05, seed = NULL, sig2 = 0.3, shuffle = FALSE,
 #' hrm.freqs = c(1/2, 1/3, 2/3, 1/4, 3/4, 1/5, 1/6, 1/7, 1/8, 1/9, 1/10),
-#' sampling.rate = 44.1)
+#' sampling.rate = 44.1, pb = TRUE)
 #' @param replicates Numeric vector of length 1 indicating the number of replicates for each treatment combination. Default is 1. Useful for measuring variation in transmission parameters.
 #' @param frequencies Numeric vector with the different frequencies (in seconds) to synthesize. A Brownian bridge motion stochastic process (\code{diff.fun == "BB"}) is used to simulate frequency modulation (see \code{\link[warbleR]{simulate_songs}}).
 #' @param durations Numeric vector with the different durations (in seconds) to synthesize.
@@ -21,6 +21,7 @@
 #' @param shuffle Logical to control if the position of sounds is randomized. Having all sounds from the same treatment in a sequence can be problematic if an environmental noise masks them. Hence 'shuffle' is useful to avoid having sounds from the same treatment next to each other. Default is \code{FALSE}.
 #' @param hrm.freqs Numeric vector with the frequencies of the harmonics relative to the fundamental frequency. The default values are c(1/2, 1/3, 2/3, 1/4, 3/4, 1/5, 1/6, 1/7, 1/8, 1/9, 1/10).
 #' @param sampling.rate Numeric vector of length 1. Sets the sampling frequency of the wave object (in kHz). Default is 44.1.
+#' @param pb Logical argument to control if progress bar is shown. Default is \code{TRUE}.
 #' @return An extended selection table, which can be input into \code{\link{master_sound_file}} to create the .wav file. The table contains columns for each of the varying features a 'treatment' column (useful to tell the acoustic features of each sound) and a 'replicate' column indicating the replicates for each 'treatment'.
 #' @family prepare acoustic data
 #' @seealso \code{\link[warbleR]{simulate_songs}} from the package warbleR.
@@ -70,7 +71,8 @@ synth_sounds <-
            shuffle = FALSE,
            hrm.freqs = c(1 / 2, 1 / 3, 2 / 3, 1 / 4, 3 / 4, 1 / 5, 1 / 6, 1 / 7, 1 /
                            8, 1 / 9, 1 / 10),
-           sampling.rate = 44.1) {
+           sampling.rate = 44.1,
+           pb = TRUE) {
     if (am & length(am.amps) < 2) {
       stop2("if 'am = TRUE' 'am.amps' must have more than 1 value 'length(am.amps) > 1'")
     }
@@ -111,7 +113,7 @@ synth_sounds <-
     
     # simulate songs
     sim.songs <-
-      warbleR:::pblapply_wrblr_int(seq_len(nrow(eg)), function(x) {
+      warbleR:::pblapply_wrblr_int(seq_len(nrow(eg)), pbar = pb, function(x) {
         sm.sng <- warbleR::simulate_songs(
           n = length(frequencies),
           durs = eg$dur[x],
