+++ .github/workflows/deploy_ghpages.yaml
@@ -2,54 +2,54 @@ name: Continuous Deployment
 
 on:
   push:
-  branches:
-  - master
-schedule:
-  - cron: '0 0 * * *'
+    branches:
+      - master
+  schedule:
+    - cron: '0 0 * * *'
 
 jobs:
   build-and-deploy:
-  name: Build and deploy to Github Pages
-runs-on: ubuntu-latest
-steps:
-  - name: Checkout repo
-uses: actions/checkout@v3
-- name: Use nodejs
-uses: actions/setup-node@v3
-with:
-  node-version: '16.x'
-- name: Get yarn cache directory path
-id: yarn-cache-dir-path
-run: echo "::set-output name=dir::$(yarn cache dir)"
-- name: Activate dependency cache
-uses: actions/cache@v3
-id: yarn-cache # use this to check for `cache-hit` (`steps.yarn-cache.outputs.cache-hit != 'true'`)
-with:
-  path: ${{ steps.yarn-cache-dir-path.outputs.dir }}
-key: ${{ runner.os }}-yarn-${{ hashFiles('**/yarn.lock') }}
-restore-keys: |
-  ${{ runner.os }}-yarn-
-  - name: Install dependencies
-run: yarn install --frozen-lockfile
-- name: Caching Gatsby
-id: gatsby-cache-build
-uses: actions/cache@v3
-with:
-  path: |
-  public
-.cache
-key: ${{ runner.os }}-gatsby-build-cache-${{ github.run_id }}
-restore-keys: |
-  ${{ runner.os }}-gatsby-build-cache-
-  - name: Build website
-run: yarn build:with-prefix
-env:
-  PATH_PREFIX: '/baRulho'
-SITE_URL: 'marce10.github.io/baRulho/'
-CI: true
-- name: Deploy to GitHub Pages
-uses: JamesIves/github-pages-deploy-action@v4.3.3
-with:
-  branch: gh-pages
-folder: public
-clean: true
+    name: Build and deploy to Github Pages
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout repo
+        uses: actions/checkout@v3
+      - name: Use nodejs
+        uses: actions/setup-node@v3
+        with:
+          node-version: '16.x'
+      - name: Get yarn cache directory path
+        id: yarn-cache-dir-path
+        run: echo "::set-output name=dir::$(yarn cache dir)"
+      - name: Activate dependency cache
+        uses: actions/cache@v3
+        id: yarn-cache # use this to check for `cache-hit` (`steps.yarn-cache.outputs.cache-hit != 'true'`)
+        with:
+          path: ${{ steps.yarn-cache-dir-path.outputs.dir }}
+          key: ${{ runner.os }}-yarn-${{ hashFiles('**/yarn.lock') }}
+          restore-keys: |
+            ${{ runner.os }}-yarn-
+      - name: Install dependencies
+        run: yarn install --frozen-lockfile
+      - name: Caching Gatsby
+        id: gatsby-cache-build
+        uses: actions/cache@v3
+        with:
+          path: |
+            public
+            .cache
+          key: ${{ runner.os }}-gatsby-build-cache-${{ github.run_id }}
+          restore-keys: |
+            ${{ runner.os }}-gatsby-build-cache-
+      - name: Build website
+        run: yarn build:with-prefix
+        env:
+          PATH_PREFIX: '/baRulho'
+`         SITE_URL: 'marce10.github.io/baRulho/'
+          CI: true
+      - name: Deploy to GitHub Pages
+        uses: JamesIves/github-pages-deploy-action@v4.3.3
+        with:
+          branch: gh-pages
+          folder: public
+          clean: true
