### 2023-10-03 83c5864
try updating tic.yml and other to build/deploy
