### 2023-10-03 2117e8f
Fix package documentation so that pkgdown should build
