+++ NAMESPACE
@@ -24,3 +24,41 @@ export(spectrum_blur_ratio)
 export(spectrum_correlation)
 export(synth_sounds)
 export(tail_to_signal_ratio)
+import(Sim.DiffProc)
+import(fftw)
+import(grDevices)
+import(graphics)
+import(parallel)
+import(png)
+import(seewave)
+import(tuneR)
+import(utils)
+import(viridis)
+import(warbleR)
+importFrom(checkmate,assert_character)
+importFrom(checkmate,assert_class)
+importFrom(checkmate,assert_data_frame)
+importFrom(checkmate,assert_directory)
+importFrom(checkmate,assert_function)
+importFrom(checkmate,assert_integerish)
+importFrom(checkmate,assert_list)
+importFrom(checkmate,assert_logical)
+importFrom(checkmate,assert_multi_class)
+importFrom(checkmate,assert_names)
+importFrom(checkmate,assert_numeric)
+importFrom(checkmate,makeAssertCollection)
+importFrom(checkmate,makeAssertionFunction)
+importFrom(checkmate,reportAssertions)
+importFrom(cli,make_ansi_style)
+importFrom(cli,num_ansi_colors)
+importFrom(cli,style_bold)
+importFrom(cli,style_italic)
+importFrom(methods,is)
+importFrom(ohun,template_correlator)
+importFrom(ohun,template_detector)
+importFrom(stats,aggregate)
+importFrom(stats,approx)
+importFrom(stats,ave)
+importFrom(stats,complete.cases)
+importFrom(stats,cor)
+importFrom(stats,na.omit)
