### 2023-10-03 8b939cc
Fix an incomplete EOL so that pkgdown should build
