### 2023-10-03 49d2a49
Update CITATION.cff
