### 2023-10-03 c488dfb
added new function add_noise
