### 2023-10-03 11eb01e
fix bug add_noise
