+++ pkgdown/_pkgdown.yml
@@ -52,4 +52,5 @@ reference:
 - title: Additional functions
   contents:
   - noise_profile
+  - add_noise
   
