### 2023-10-03 #609 @jhollist
---

@maRce10 See checks in previous comment.  R CMD Check is now failing.  I did see that you recently merged some new work and that is triggering this failure.  I assume you are still doing a bit of work on this.  Let me know when you have that completed and your checks pass.  I will check again and then move forward with finding reviewers.

I also noted a few others things.  Data is a bit large, not sure if you can reduce that.

```
❯ checking installed package size ... NOTE
      data   1.6Mb
      doc    7.5Mb
      help   1.5Mb
```
- I see 6 Open Issues - what are the status on those?  If closed or no longer relevant, clean those up.  If still relevant provide updates in issues so that reviewers can asses the status.
---
