+++ R/signal_to_noise_ratio.R
@@ -78,19 +78,19 @@ signal_to_noise_ratio <-
            path = getOption("sound.files.path", ".")) {
     # check arguments
     arguments <- as.list(base::match.call())
-    
+
     # add objects to argument names
     for (i in names(arguments)[-1]) {
       arguments[[i]] <- get(i)
     }
-    
+
     # check each arguments
     check_results <-
       check_arguments(fun = arguments[[1]], args = arguments)
-    
+
     # report errors
     report_assertions2(check_results)
-    
+
     # get sampling rate
     sampling_rate <-
       warbleR::read_sound_file(
@@ -99,49 +99,52 @@ signal_to_noise_ratio <-
         path = path,
         header = TRUE
       )$sample.rate
-    
+
     # adjust wl based on hop.size
     if (is.null(wl)) {
-      wl <- round(sampling_rate * hop.size / 1000,
-                  0)
+      wl <- round(
+        sampling_rate * hop.size / 1000,
+        0
+      )
     }
-    
+
     # make wl even if odd
     if (!(wl %% 2) == 0) {
       wl <- wl + 1
     }
-    
+
     # check sound.id column
     if (is.null(X$sound.id)) {
       if (noise.ref == "custom") {
         stop2("'sound.id' required when 'noise.ref == 'custom''")
       }
-      
+
       X$sound.id <- "no.sound.id.column"
     }
-    
+
     # check if 'ambient' is found in  sound.id column
     if (!any(X$sound.id %in% "ambient") &
-        noise.ref == "custom") {
+      noise.ref == "custom") {
       stop2(
         "'ambient' selections must be contained in 'X' and labeled in a 'sound.id' column as 'ambient' when 'noise.ref = 'custom'"
       )
     }
-    
+
     if (noise.ref == "custom" &
-        any(vapply(unique(X$sound.files), function(x) {
-          sum(X$sound.files == x &
-              X$sound.id == "ambient")
-        }, FUN.VALUE = numeric(1)) == 0)) {
+      any(vapply(unique(X$sound.files), function(x) {
+        sum(X$sound.files == x &
+          X$sound.id == "ambient")
+      }, FUN.VALUE = numeric(1)) == 0)) {
       stop2(
         "Each sound file referenced in 'X' must have at least 1 'ambient' selection when 'noise.ref = 'custom'"
       )
     }
-    
+
     # 'mar' is needed when not using equal duration
-    if (!eq.dur & is.null(mar))
-        stop2("'mar' must be supplied when 'eq.dur = FALSE'")
-      
+    if (!eq.dur & is.null(mar)) {
+      stop2("'mar' must be supplied when 'eq.dur = FALSE'")
+    }
+
     # set clusters for windows OS
     if (Sys.info()[1] == "Windows" & cores > 1) {
       cl <-
@@ -149,7 +152,7 @@ signal_to_noise_ratio <-
     } else {
       cl <- cores
     }
-    
+
     # calculate all RMS of envelopes with a apply function
     rms_list <-
       warbleR:::pblapply_wrblr_int(
@@ -157,24 +160,24 @@ signal_to_noise_ratio <-
         pbar = pb,
         cl = cl,
         FUN = function(y) {
-          
           # only calculate for non-markers and for ambient only if custom noise.ref
-          if (!X$sound.id[y] %in% c("marker", if (noise.ref != "custom") "ambient")){
-            
+          if (!X$sound.id[y] %in% c("marker", if (noise.ref != "custom") "ambient")) {
             if (noise.ref == "custom") {
               # read sound clip
               signal <-
-                warbleR::read_sound_file(X = X,
-                                         index = y,
-                                         path = path)
-              
+                warbleR::read_sound_file(
+                  X = X,
+                  index = y,
[…]
