+++ R/internal_functions.R
@@ -22,7 +22,7 @@ find_peaks_bRlh_int <-
     } else {
       cl <- cores
     }
-    
+
     # loop over scores of each dyad
     pks <-
       warbleR:::pblapply_wrblr_int(
@@ -32,69 +32,70 @@ find_peaks_bRlh_int <-
         FUN = function(i) {
           # extract data for a dyad
           dat <- xc.output$scores[xc.output$scores$dyad == i, ]
-          
+
           # check xc.output being a autodetec.output object
           if (!(is(xc.output, "xcorr.output") |
-                is(xc.output, "xc.output"))) {
+            is(xc.output, "xc.output"))) {
             stop2("'xc.output' must be and object of class 'xcorr.output'")
           }
-          
+
           ## get peaks as the ones higher than previous and following scores
           pks <-
             dat[c(FALSE, diff(dat$score) > 0) &
-                  c(rev(diff(rev(dat$score)) > 0), FALSE) &
-                  dat$score > cutoff, , drop = FALSE]
-          
+              c(rev(diff(rev(dat$score)) > 0), FALSE) &
+              dat$score > cutoff, , drop = FALSE]
+
           # get the single highest peak
           if (max.peak) {
             pks <- dat[which.max(dat$score), , drop = FALSE]
           }
-          
+
           return(pks)
         }
       )
-    
+
     # put results in a data frame
     peaks <- do.call(rbind, pks)
-    
+
     # relabel rows
     if (nrow(peaks) > 0) {
       rownames(peaks) <- seq_len(nrow(peaks))
-      
+
       # remove dyad column
       peaks$dyad <- NULL
-      
+
       #### name as in a warbleR selection table
       # remove selec info at the end
       peaks$sound.files <-
         substr(peaks$sound.files,
-               start = 0,
-               regexpr("\\-[^\\-]*$", peaks$sound.files) - 1)
-      
+          start = 0,
+          regexpr("\\-[^\\-]*$", peaks$sound.files) - 1
+        )
+
       #### add start and end
       # add template column to selection table in xc.output
       Y <- xc.output$org.selection.table
       Y$template <- paste(Y$sound.files, Y$selec, sep = "-")
-      
+
       # Y <- Y[Y$template %in% comp_mat[, 1], ]
-      
+
       # add start as time - half duration of template
       peaks$start <- vapply(seq_len(nrow(peaks)), function(i) {
         peaks$time[i] -
           ((Y$end[Y$template == peaks$template[i]] -
-              Y$start[Y$template == peaks$template[i]]) / 2)
+            Y$start[Y$template == peaks$template[i]]) / 2)
       }, FUN.VALUE = numeric(1))
-      
+
       # add end as time + half duration of template
       peaks$end <- vapply(seq_len(nrow(peaks)), function(i) {
         peaks$time[i] +
           ((Y$end[Y$template == peaks$template[i]] -
-              Y$start[Y$template == peaks$template[i]]) / 2)
+            Y$start[Y$template == peaks$template[i]]) / 2)
       }, FUN.VALUE = numeric(1))
-      
+
       # add selec labels
       peaks$selec <- 1
-      
+
       if (nrow(peaks) > 1) {
         for (i in 2:nrow(peaks)) {
           if (peaks$sound.files[i] == peaks$sound.files[i - 1]) {
@@ -102,10 +103,10 @@ find_peaks_bRlh_int <-
           }
         }
       }
-      
+
       # sort columns in a intuitive order
[…]
