### 2023-10-03 804ecc7
Update add_noise.Rd
