+++ man/add_noise.Rd
@@ -37,7 +37,14 @@ The function adds synthetic noise to sounds referenced in an extended selection 
 }
 \examples{
 \dontrun{
+# load example data
+data("test_sounds_est")
 
+# make it a 'by song' extended selection table
+X <- baRulho:::by_element_est(X = test_sounds_est)
+
+# add noise to the first five rows
+X_noise <- add_noise(X = X[1:5, ], mar = 0.2, target.snr = 3)
 }
 }
 \references{
