+++ NAMESPACE
@@ -1,6 +1,5 @@
 # Generated by roxygen2: do not edit by hand
 
-export(add_noise)
 export(align_test_files)
 export(attenuation)
 export(auto_realign)
