### 2023-10-04 a48386a
Update quantify_degradation.Rmd
