### 2023-10-04 33c399e
Add depends to vignette and align index entry with title
