### 2023-10-04 ff93d36
Remove spurious function that doesn't exist preventing pkgdown from building site
