+++ pkgdown/_pkgdown.yml
@@ -52,4 +52,3 @@ reference:
 - title: Additional functions
   contents:
   - noise_profile
-  - add_noise
