+++ pkgdown/_pkgdown.yml
@@ -51,4 +51,5 @@ reference:
 
 - title: Additional functions
   contents:
+  - add_noise
   - noise_profile
