+++ R/add_noise.R
@@ -1,0 +1,201 @@
+#' Add synthetic noise
+#'
+#' \code{add_noise} adds synthetic noise to annotations in extended selection tables
+#' @usage add_noise(X, mar = NULL, target.snr = 2, precision = 0.1,
+#' cores = getOption("mc.cores", 1), pb = getOption("pb", TRUE),
+#' max.iterations = 1000, ...)
+#' @param X Object of class 'extended_selection_table' (created by the function \code{\link[warbleR]{selection_table}} from the warbleR package), generated 'by element', with the reference to the test sounds (typically the output of \code{\link{align_test_files}}). Must contain the following columns: 1) "sound.files": name of the .wav files, 2) "selec": unique selection identifier (within a sound file), 3) "start": start time and 4) "end": end time of selections, 5) "bottom.freq": low frequency for bandpass, 6) "top.freq": high frequency for bandpass and 7) "sound.id": ID of sounds used to identify counterparts across distances (only needed for "custom" noise reference, see "noise.ref" argument).
+#' @param mar numeric vector of length 1. Specifies the margins adjacent to
+#'   the start point of the annotation over which to measure ambient noise.
+#' @param target.snr numeric vector of length 1. Specifies the desired signal-to-noise ratio. Must be lower that the current signal-to-noise ratio. Annotations showing a signal-to-noise ratio higher than 'target.snr' will remain unchanged. Must be supplied.
+#' @param precision  numeric vector of length 1. Specifies the precision of the adjusted signal-to-noise ratio (in dB).
+#' @param cores Numeric vector of length 1. Controls whether parallel computing is applied by specifying the number of cores to be used. Default is 1 (i.e. no parallel computing).
+#' @param pb Logical argument to control if progress bar is shown. Default is \code{TRUE}.
+#' @param max.iterations Numeric vector of length 1. Specifies the maximum number of iterations that the internal signal-to-noise adjusting routine will run before stopping. Note that in most cases the default maximum number of iterations (1000) is not reached.
+#' @param ... Additional arguments to be passed internally to \code{\link{signal_to_noise_ratio}}.
+#' @return Object 'X' in which the wave objects have been modified to match the target signal-to-noise ratio. It also includes an additional column, 'adjusted.snr', with the new signal-to-noise ratio values.
+#' @export
+#' @name add_noise
+#' @details The function adds synthetic noise to sounds referenced in an extended selection table (class created by the function \code{\link[warbleR]{selection_table}} from the warbleR package) to decrease the signal-to-noise ratio. This can be useful, for instance, for evaluating the effect of background noise on signal structure. Note that the implementation is slow.
+#' @examples \dontrun{
+#' # load example data
+#' data("test_sounds_est")
+#'
+#' # make it a 'by song' extended selection table
+#' X <- baRulho:::by_element_est(X = test_sounds_est)
+#'
+#' # add noise to the first five rows
+#' X_noise <- add_noise(X = X[1:5, ], mar = 0.2, target.snr = 3)
+#' }
+#' @author Marcelo Araya-Salas (\email{marcelo.araya@@ucr.ac.cr})
+#' @family miscellaneous
+#' @seealso \code{\link{signal_to_noise_ratio}}
+#' @references {
+#' Araya-Salas, M. (2020). baRulho: baRulho: quantifying degradation of (animal) acoustic signals in R. R package version 1.0.2
+#' }
+
+add_noise <-
+  function(X,
+           mar = NULL,
+           target.snr = 2,
+           precision = 0.1,
+           cores = getOption("mc.cores", 1),
+           pb = getOption("pb", TRUE),
+           max.iterations = 1000,
+           ...) {
+    # check arguments
[…]
