### 2023-10-04 145c2c0
switch attenuation to miscellaneous functions
