+++ man/attenuation.Rd
@@ -42,20 +42,11 @@ Araya-Salas, M. (2020). baRulho: baRulho: quantifying degradation of (animal) ac
 }
 }
 \seealso{
-Other quantify degradation: 
-\code{\link{blur_ratio}()},
-\code{\link{detection_distance}()},
-\code{\link{envelope_correlation}()},
-\code{\link{plot_blur_ratio}()},
-\code{\link{plot_degradation}()},
-\code{\link{set_reference_sounds}()},
-\code{\link{signal_to_noise_ratio}()},
-\code{\link{spcc}()},
-\code{\link{spectrum_blur_ratio}()},
-\code{\link{spectrum_correlation}()},
-\code{\link{tail_to_signal_ratio}()}
+Other miscellaneous: 
+\code{\link{add_noise}()},
+\code{\link{noise_profile}()}
 }
 \author{
 Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
 }
-\concept{quantify degradation}
+\concept{miscellaneous}
