+++ man/envelope_correlation.Rd
@@ -74,7 +74,6 @@ Apol, C.A., Sturdy, C.B. & Proppe, D.S. (2017). Seasonal variability in habitat 
 \code{\link{blur_ratio}}, \code{\link{spectrum_blur_ratio}}
 
 Other quantify degradation: 
-\code{\link{attenuation}()},
 \code{\link{blur_ratio}()},
 \code{\link{detection_distance}()},
 \code{\link{plot_blur_ratio}()},
