### 2023-10-04 2cf6a40
Deploy from Github Actions build 6410111281 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/6410111281
Commit: 145c2c00c6783d64f2b4f9030ad1b962cf32c699
