+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   align_test_sounds: align_test_sounds.html
   quantify_degradation: quantify_degradation.html
-last_built: 2023-10-04T18:28Z
+last_built: 2023-10-04T18:46Z
 urls:
   reference: https://marce10.github.io/baRulho/reference
   article: https://marce10.github.io/baRulho/articles
