### 2023-10-04 36f9024
fix bug when power spectrum cannot be computed
