+++ R/synth_sounds.R
@@ -57,6 +57,7 @@
 #' }
 #' @author Marcelo Araya-Salas (\email{marcelo.araya@@ucr.ac.cr})
 
+
 synth_sounds <-
   function(replicates = 1,
            frequencies,
@@ -70,21 +71,36 @@ synth_sounds <-
            sig2 = 0.3,
            shuffle = FALSE,
            hrm.freqs = c(1 / 2, 1 / 3, 2 / 3, 1 / 4, 3 / 4, 1 / 5, 1 / 6, 1 / 7, 1 /
-             8, 1 / 9, 1 / 10),
+                           8, 1 / 9, 1 / 10),
            sampling.rate = 44.1,
            pb = TRUE) {
+    # check arguments
+    arguments <- as.list(base::match.call())
+    
+    # add objects to argument names
+    for (i in names(arguments)[-1]) {
+      arguments[[i]] <- get(i)
+    }
+    
+    # check each arguments
+    check_results <-
+      check_arguments(fun = arguments[[1]], args = arguments)
+    
+    # report errors
+    report_assertions2(check_results)
+    
+    
     if (am & length(am.amps) < 2) {
       stop2("if 'am = TRUE' 'am.amps' must have more than 1 value 'length(am.amps) > 1'")
     }
-
+    
     # set number of steps (default is 10)
     steps <- if (length(am.amps) > 2) {
       length(am.amps)
     } else {
       10
     }
-
-
+    
     # make all possible combinations
     eg <- expand.grid(
       duration = durations,
@@ -93,24 +109,24 @@ synth_sounds <-
       harmonics = c("no.harmonics", "harmonics"),
       stringsAsFactors = FALSE
     )
-
+    
     # remove levels when treatments not included
     if (!fm) {
-      eg <- eg[eg$fm == "no.fm", ]
+      eg <- eg[eg$fm == "no.fm",]
     }
-
+    
     if (!am) {
-      eg <- eg[eg$am == "no.am", ]
+      eg <- eg[eg$am == "no.am",]
     }
-
+    
     if (nharmonics == 1) {
-      eg <- eg[eg$harmonics == "no.harmonics", ]
+      eg <- eg[eg$harmonics == "no.harmonics",]
     }
-
+    
     # create temporary directory
     temp_dir <- tempfile()
     dir.create(temp_dir)
-
+    
     # simulate songs
     sim.songs <-
       warbleR:::pblapply_wrblr_int(seq_len(nrow(eg)), pbar = pb, function(x) {
@@ -143,13 +159,13 @@ synth_sounds <-
           selec.table = TRUE,
           sig2 = sig2,
           steps = steps,
-          file.name = paste(eg[x, ], collapse = "_"),
+          file.name = paste(eg[x,], collapse = "_"),
           bgn = 0,
           seed = seed,
           path = temp_dir,
           hrm.freqs = hrm.freqs
         )
-
+        
         # add freq room if pure tone
         if (eg$fm[x] == "no.fm") {
           sm.sng$selec.table$bottom.freq <-
@@ -157,33 +173,33 @@ synth_sounds <-
           sm.sng$selec.table$top.freq <-
             sm.sng$selec.table$top.freq + 0.2
         }
-
+        
         sm.sng$selec.table$bottom.freq[sm.sng$selec.table$bottom.freq < 0] <-
           0.1
-
+        
         sm.sng$selec.table$sim.freq <- as.character(frequencies)
-
+        
         return(sm.sng)
       })
-
-
+    
+    
     # name with parameters
     names(sim.songs) <-
       vapply(seq_len(nrow(eg)), function(x) {
-        paste(eg[x, ], collapse = "_")
+        paste(eg[x,], collapse = "_")
       }, FUN.VALUE = character(1))
-
+    
[…]
