### 2023-10-04 #609 @maRce10
OK I solved and closed the issues and now R CMD check is working again. The size of data is probably due to having sounds as examples. I shrank it as much as I could.
