### 2023-10-05 #609 @jhollist
@maRce10 I think we are good to move on to review.

Couple of things to think about in the lintr results, in particular the 1:nrow and sapply ones.  The library calls you can ignore since those are in Rmd and tests.

Look through those results for other ideas as well.  For instance, your lines of code are on the high side, but so is cyclomatic complexity.  If you can simplify some of these functions you could probably reduce both and make future maintenance a little easier.

That being said, we are good to go for reviewers.  I'll start pinging on a few folks today.
