### 2023-10-10 9cb7b0b
Deploy from Github Actions build 6472054539 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/6472054539
Commit: 3ad91151c3c723493113b3c8035abc6427125acc
