### 2023-10-14 b244c04
Deploy from Github Actions build 6515497910 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/6515497910
Commit: 3ad91151c3c723493113b3c8035abc6427125acc
