### 2023-10-15 2c699c2
Deploy from Github Actions build 6521898010 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/6521898010
Commit: 3ad91151c3c723493113b3c8035abc6427125acc
