### 2023-10-20 12a1fa1
Deploy from Github Actions build 6583296718 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/6583296718
Commit: 3ad91151c3c723493113b3c8035abc6427125acc
