### 2023-10-30 #609 @mikemahoney218
- Coverage appears to be 79.72%, which is a bit on the low side, but that's caused in large part due to `manual_realign.R` having 0% coverage. I don't know if it's possible to add automated tests for a manually interactive function, but if any of the components can be split out and tested independently that would help increase coverage. But I'm happy to check this box now, anyways.

---

The baRulho package is a useful extension to what appears to be a developing universe of packages for acoustic experiments, building on top of the ohun and warbleR packages by the same author. The package is very interesting and, while this is not my area of practice, it seems to me that this package will be very useful for people working in this space. I also want to specifically call out `manual_realign()` as being _incredibly_ cool.

My main general comment about the package is, as Jeff and the automatic checks both identified, many functions in this package have a pretty high cyclomatic complexity. As an outside reviewer, this made it hard to trace any given function. I'd encourage the developers to break their functions down into more modular pieces, in order to possibly be able to reuse elements across functions but also to make the package easier to scan and maintain in the long run.

Below I've got more specific comments, mostly focused on vignettes and other documentation. Overall, I think most of these comments are fiddly details and none are fundamental problems with the package. I will say that this package is not within my area of study -- Jeff asked me to review "for R package fundamentals and general ecology applicability" and that's where my comments are focused. That means that I am not capable of assessing if any of the functions or equations used are _correct_, and so I just assumed that the approaches used were appropriate.

#### Check problems

Running R CMD check, I see:

```
❯ checking installed package size ... NOTE
      data   1.6Mb
      doc    8.8Mb
      help   1.5Mb
```

I saw the earlier conversation between you and Jeff on this, but I think the
size isn't due to using sound examples -- I think it's due to the size of images
you're using in your help files and vignettes. It seems like the PNGs in the
package are saved at 300 PPI resolution; saving them at 72 or 96 PPI instead
will save a ton of space (for instance, it shrinks the package sticker from 1.4
MB to 162.5 kB on my machine) while leaving them high-res enough for computer
displays.

You might also consider moving the quantify_degradation vignette to be an
"article" instead, saving some of the space that gets used for those images.

##### align_test_sounds.Rmd

In general, I'd encourage you to evaluate all chunks in a vignette unless
they're expected to error, on purpose, and this is highlighted to the reader.
I get into specifics below, but reading the rendered vignette I am confused by
what some functions are meant to output, and I believe there are some code
errors hidden in unevaluated chunks.

This vignette defines `td` in a setup chunk that's not included in the vignette,
then uses it in user-visible code throughout the vignette. As a user, I'd expect
to be able to run the code displayed in the vignette. I'd suggest either
defining `td` in the user-visible part of the vignette, or using `tempdir()` as
an argument to these functions, rather than using the variable.

The calls to `spectrograms()` and `check_sels()` using `synth_master_annotations`
error ("`NAs found in bottom.freq and/or top.freq`"). Those chunks are marked
as `eval = FALSE`, so maybe this is known, but there's nothing in the vignette
implying that these chunks will error -- in fact, there are spectrograms
included below the chunk, which mean I expect that running the code in that
chunk will produce the resulting graph.

Is there any reason you switch from using arguments to set file paths to using
options right before `find_markers()`? The inconsistency confused me.

You have a hidden call to `print(TRUE)` after calling
`is_extended_selection_table()`. I think it would probably be better to return
the actual output from `is_extended_selection_table()` instead. On my machine,
this returned `FALSE`, making the vignette incorrect (the actual class is
`data.frame`).

The command `\code{\link{align_test_files}}` in the manually fixing alignment
section produces a blank space in the HTML output. I think this should probably
be `align_test_files()`.

In `auto_realign()` I'm seeing:

```
1: In attr(X, "check.results")$start[paste(X$sound.files, X$selec,  :
  number of items to replace is not a multiple of replacement length
2: In attr(X, "check.results")$end[paste(X$sound.files, X$selec, sep = "-") %in%  :
  number of items to replace is not a multiple of replacement length
```

Is that expected?

##### quantify_degradation.Rmd

- `options(dest.path = td)` (in the blur ratio section) errors because td isn't
  defined. As with the other vignette, the mix of hidden evaluated code and
  unevaluated displayed code means the actual code displayed in the vignette
  doesn't always work.

- There's a line in this vignette that says "The following code sets env.smooth = 800", but it's not obvious to me where or why that gets set.

- Some of the plots use `scale_color_viridis()`, some use `scale_color_manual(values = viridis::viridis())`, and it's not obvious to me why this changes. It makes comparing the code in different chunks hard, as there are a number of things changing for reasons I can't identify.

- I don't know if it's your CSS formatting or something else, but while code in
  the Rmd is formatted nicely, the ggplot2 code on the pkgdown site seems to all
  be on a single line in a way that makes it hard to parse.

#### Examples

If you're going to use `\dontrun` around examples that should work, which don't
highlight to users that they're expected to fail, I recommend using
`devtools::run_examples(run_donttest = TRUE, run_dontrun = TRUE)` to confirm
that the code actually will run successfully.

I see the following errors:

- blur_ratio.R: need to use `ggplot2::` or `library(ggplot2)`. It's probably
  good practice to also wrap this in
  `if (requireNamespace("ggplot2", quietly = TRUE))` (or, if you're willing to
  pick up a dependency on rlang, `rlang::check_installed("ggplot2")`).
- manual_realign.R: `Error: object 'master_df' not found`. Seems like `Y` should
  be `as.data.frame(master_est)` (or this object needs to be created earlier)?
- manual_realign.R: `Error in eval(ei, envir) : object 'mako' not found`. Need
  to namespace `viridis::mako()`.
- spectrum_blur_ratio.R: also needs `ggplot2::` or `library(ggplot2)`.

Running examples, I also see:

```
1: In attenuation(f = 2000, dist = 50, dist0 = 1) :
  partial argument match of 'f' to 'frequency'
2: In match.call(definition, call, expand.dots, envir) :
  partial argument match of 'f' to 'frequency'
3: In attr(X, "check.results")$start[paste(X$sound.files, X$selec,  :
  number of items to replace is not a multiple of replacement length
4: In attr(X, "check.results")$end[paste(X$sound.files, X$selec, sep = "-") %in%  :
  number of items to replace is not a multiple of replacement length
5: In eg$dur : partial match of 'dur' to 'duration'
6: In eg$harm : partial match of 'harm' to 'harmonics'
7: In eg$harm : partial match of 'harm' to 'harmonics'
```

- blur_ratio.R: Any reason the call to `blur_ratio()` on line 42 is commented out?

#### Package functions

Is there a reason you're using the `@usage` tag? Normally this section is
auto-generated by rOxygen2, ensuring it stays up-to-date and properly formatted.

I love the standardization of arguments across functions; I think that's a clear
strength of the package. You might consider using rOxygen2 templates (see the
[relevant section of the packaging guide](https://devguide.ropensci.org/building.html#roxygen2-use))
to centralize these arguments' documentation, in order to reduce the amount of
places you need to touch to update any documentation.

- It might be nice to define more of the constants used in `attenuation()`, so
  that the calculations are a bit easier to interpret. Right now, there's a lot
  of [magic numbers](https://en.wikipedia.org/wiki/Magic_number_(programming))
  in this function call.

- In `attenuation()`, I think `dist0` and `dist` are required inputs, but are
  placed after optional arguments. It might be nicer to place these arguments
  after `frequency`, to allow users to pass the distance arguments via position.

- I think this is probably the easiest time to get rid of deprecated arguments
  and functions, before the package gets to CRAN/gets more visibility via the
  rOpenSci organization. I see the `parallel` argument and `output` argument in
  several functions is deprecated, and I think this is a good time to drop
  those, as well as the `search_templates()` and `spcc_distortion()` functions.

- `auto_realign()`: I'd recommend providing the list of options to `wn` here.
  See the [Enumerate possible options](https://design.tidyverse.org/enumerate-options.html)
  in the "Tidy design principles" book for an example of why this is useful.

- I'd similarly suggest using enumerations for `noise.ref`, `type`, `method`,
  and other functions where users must provide an option from a pre-defined
  list. Then use `arg.match()` or `rlang::match_arg()` to ensure provided values
  match one of the options.

- `blur_ratio()`, `plot_aligned_sounds()`: I think it'd make sense to mention in
  the help file that `wl` must be even, and that it will be adjusted if not.
  Also, does `wl` need to be
  an integer value? It's documented as a numeric vector, but any decimal values
  will have 1 added to them.

- `set_reference_sounds()`: I see that this function requires eight columns with
  specific names. Is there any reason to not let the user control, via function
  arguments, which columns are used?

- Consider having gunctions that generate plots (I'm thinking of
  `plot_degradation()`, `plot_blur_ratio()`) return the file paths they
  generated. I was surprised a few times when a function appeared to not return
  anything, because it had written a figure out to disk instead.

- `plot_blur_ratios()`: `colors` says it should be a `Character vector of length 4`,
  but then the default argument is `viridis::viridis(3)`, of length 3. Not sure
  what's correct here.

- `tail_to_signal_ratio()`: `bp` having a default value that doesn't have the
  same type or length as what users would provide is a bit confusing. Maybe this
  default should be `NULL`?.

- `type` seems to be used in different ways in several different functions,
  with different sets of options in each. Consider using different argument
  names for each.

#### Misc

The contributing guide seems to be missing an email:

> You must never report security related issues, vulnerabilities or bugs including sensitive information to the issue tracker, or elsewhere in public. Instead sensitive bugs must be sent by email to <>.

I'd either provide an email or drop this part of the guide.
