### 2023-11-04 196715d
Deploy from Github Actions build 6752695854 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/6752695854
Commit: 3ad91151c3c723493113b3c8035abc6427125acc
