### 2023-11-04 981f60b
added check fir exrended selection tables created by element
