+++ vignettes/align_test_sounds.Rmd
@@ -511,7 +511,10 @@ manual_realign(X = aligned_tests, Y = master_annotations, duration = 2.4)
 
 ```
 
- <img src="manual_realign.gif" alt="manual realign">
+```{r, echo=FALSE, out.width="100%", fig.align='center'}
+knitr::include_graphics("manual_realign.gif")
+```
+ <!-- <img src="manual_realign.gif" alt="manual realign"> -->
 
 Spectrograms include the first few seconds of the sound files (controlled by 'duration') which is usually enough to tell the precision of the alignment. The lower spectrogram shows a series of 'buttons' that users can click on to control if the test sound file spectrogram (low panel) needs to be moved to the left ("<") or right (">"). Users can also reset the spectrogram to its original position ('reset'), move on to the next sound file in 'X' (test sound file annotations) or stop the process (stop button). The function returns an object similar to the input object 'X' in which the start and end of the sounds have been adjusted. By default it uses the start marker as reference but this can be modified with the argument 'marker'.
 
