### 2023-11-04 9c9d95d
reduce size sticker
