### 2023-11-06 5c6ab70
Deploy from Github Actions build 6766220487 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/6766220487
Commit: 3ad91151c3c723493113b3c8035abc6427125acc
