### 2023-11-07 15d420c
Deploy from Github Actions build 6779914094 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/6779914094
Commit: 3ad91151c3c723493113b3c8035abc6427125acc
