### 2023-11-08 e0e343b
Deploy from Github Actions build 6793678020 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/6793678020
Commit: 3ad91151c3c723493113b3c8035abc6427125acc
