### 2023-11-12 806656a
Deploy from Github Actions build 6838519465 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/6838519465
Commit: 3ad91151c3c723493113b3c8035abc6427125acc
