### 2023-11-14 ea4230a
Deploy from Github Actions build 6859122996 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/6859122996
Commit: 3ad91151c3c723493113b3c8035abc6427125acc
