### 2023-11-20 c6fc355
Deploy from Github Actions build 6925450154 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/6925450154
Commit: 3ad91151c3c723493113b3c8035abc6427125acc
