### 2023-11-21 98fa1ca
Deploy from Github Actions build 6939169839 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/6939169839
Commit: 3ad91151c3c723493113b3c8035abc6427125acc
