### 2023-11-25 71560ac
Deploy from Github Actions build 6986952848 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/6986952848
Commit: 3ad91151c3c723493113b3c8035abc6427125acc
