### 2023-12-27 b6868c4
lower cyclomatic complexity below 15
all functions now have cyclomatic complexity < 15
