+++ R/manual_realign.R
@@ -5,8 +5,8 @@
 #' wl = getOption("wl", NULL), ovlp = getOption("ovlp", 0),
 #' path = getOption("sound.files.path", "."), collevels = seq(-120, 0, 5),
 #' palette = viridis::viridis, duration = 2, mar = 0.2, step.lengths = c(5, 30),
-#' flim = NULL, label.col = "white",  ext.window = TRUE, width = 10, 
-#' height = 5, srt = 0, cex = 1, fast.spec = TRUE, 
+#' flim = NULL, label.col = "white",  ext.window = TRUE, width = 10,
+#' height = 5, srt = 0, cex = 1, fast.spec = TRUE,
 #' marker = "start_marker", grid = 0.2, ...)
 #' @param X Object of class 'data.frame', 'selection_table' or 'extended_selection_table' (the last 2 classes are created by the function \code{\link[warbleR]{selection_table}} from the warbleR package) with the reference to the test sounds (typically the output of \code{\link{align_test_files}}). Must contain the following columns: 1) "sound.files": name of the .wav files, 2) "selec": unique selection identifier (within a sound file), 3) "start": start time and 4) "end": end time of selections, 5)  "bottom.freq": low frequency for bandpass, 6) "top.freq": high frequency for bandpass and 7) "sound.id": ID of sounds used to identify counterparts across distances. Each sound must have a unique ID within a distance.
 #' @param Y object of class 'data.frame', 'selection_table' or 'extended_selection_table' (the last 2 classes are created by the function \code{\link[warbleR]{selection_table}} from the warbleR package) with the master sound file annotations. This should be the same data than that was used for finding the position of markers in \code{\link{find_markers}}. It should also contain a 'sound.id' column.
@@ -28,7 +28,7 @@
 #' @param cex Numeric argument of length 1controlling the size of sound id text labels. Default is 1.
 #' @param fast.spec Logical. If \code{TRUE} then image function is used internally to create spectrograms, which substantially increases performance (much faster), although some options become unavailable, as collevels (amplitude scale). Default is \code{FALSE}.
 #' @param marker Character string with the name of the marker to be used as the main reference for checking/adjusting time alignments. Default is 'start_marker'. Note that this can take any of the sound IDs in 'Y$sound.id'.
-#' @param grid Numeric vector of length 1 controlling the spacing between vertical lines on the spectrogram. Default is 0.2 s. Use 0 to remove grid. 
+#' @param grid Numeric vector of length 1 controlling the spacing between vertical lines on the spectrogram. Default is 0.2 s. Use 0 to remove grid.
 #' @param ... Additional arguments to be passed to the internal spectrogram
 #' creating function for customizing graphical output. The function is a modified
 #' version of \code{\link[seewave]{spectro}}, so it takes the same arguments.
@@ -56,7 +56,7 @@
 #'
 #'   # save master file
 #'   writeWave(object = attr(master_est, "wave.objects")[[1]], file.path(td, "master.wav"))
-#' 
+#'
 #'   # set path and no progress bar in global options
 #'   options(sound.files.path = td, pb = FALSE)
 #'
@@ -65,15 +65,15 @@
 #'
 #'   # align all test sounds
 #'   alg.tests <- align_test_files(X = master_est, Y = markers, pb = FALSE)
-#'   
+#'
 #'   # add error to alignment
[…]
