+++ R/plot_blur_ratio.R
@@ -73,52 +73,36 @@ plot_blur_ratio <-
            n.samples = 100) {
     # check arguments
     arguments <- as.list(base::match.call())
-
+    
     # add objects to argument names
     for (i in names(arguments)[-1]) {
       arguments[[i]] <- get(i)
     }
-
+    
     # check each arguments
     check_results <-
-      check_arguments(fun = arguments[[1]], args = arguments)
-
+      .check_arguments(fun = arguments[[1]], args = arguments)
+    
     # report errors
-    report_assertions2(check_results)
-
+    .report_assertions(check_results)
+    
     # adjust wl based on hop.size
-    if (is.null(wl)) {
-      wl <-
-        round(
-          read_sound_file(
-            X,
-            index = 1,
-            header = TRUE,
-            path = path
-          )$sample.rate * hop.size / 1000,
-          0
-        )
-    }
-
-    # make wl even if odd
-    if (!(wl %% 2) == 0) {
-      wl <- wl + 1
-    }
-
+    wl <- .adjust_wl(wl, X, hop.size)
+    
     # set clusters for windows OS
     if (Sys.info()[1] == "Windows" & cores > 1) {
       cl <- parallel::makePSOCKcluster(getOption("cl.cores", cores))
     } else {
       cl <- cores
     }
-
+    
     # add sound file selec colums to X (weird column name so it does not overwrite user columns)
     X$.sgnl.temp <- paste(X$sound.files, X$selec, sep = "-")
-
+    
     # get names of envelopes involved (those as test with reference or as reference)
     target_sgnl_temp <-
       unique(c(X$.sgnl.temp[!is.na(X$reference)], X$reference[!is.na(X$reference)]))
-
+    
     # print message
     if (pb) {
       if (type == "envelope") {
@@ -127,7 +111,7 @@ plot_blur_ratio <-
         write(file = "", x = "Computing power spectra (step 1 out of 2):")
       }
     }
-
+    
     # calculate all envelops apply function
     if (type == "envelope") {
       energy_vectors <-
@@ -142,7 +126,7 @@ plot_blur_ratio <-
                          wln = wl,
                          pth = path,
                          n.samp = n.samples) {
-            env_FUN(
+            .env(
               X = Q,
               y = x,
               env.smooth = ssmth,
@@ -154,7 +138,7 @@ plot_blur_ratio <-
           }
         )
     }
-
+    
     if (type == "spectrum") {
       # calculate all spectra apply function
       energy_vectors <-
@@ -168,7 +152,7 @@ plot_blur_ratio <-
                          Q = X,
                          pth = path,
                          n.b = n.samples) {
-            spctr_FUN(
+            .spctr(
               y,
               spec.smooth = ssmth,
               wl = wln,
@@ -179,17 +163,17 @@ plot_blur_ratio <-
           }
         )
     }
-
-
-
+    
+    
+    
     # add sound file selec column as names to envelopes
     names(energy_vectors) <- target_sgnl_temp
-
+    
     # set options to run loops
     if (pb) {
       write(file = "", x = "Producing images (step 2 out of 2):")
     }
-
+    
     # plot blur ratio
     catch_out <-
       warbleR:::pblapply_wrblr_int(
@@ -212,7 +196,7 @@ plot_blur_ratio <-
                        fl = flim,
                        cols = colors,
                        Q = X) {
-          plot_blur_FUN(
+          .plot_blur(
             X = Q,
             energy_vectors = en.vctr,
             spectr = spct,
