### 2023-12-28 26fc8d1
fix background lower panel
fix background when reaching ends lower panel
