### 2024-01-06 #609 @DenaJGibbon
## Package Review

-  [x ] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

#### Documentation

I would like to see a bit more information about the target audience for this.

- [ x] **Installation instructions:** for the development version of package and any non-standard dependencies in README

Installation was smooth.

Most of my comments for this round focused on the vignette and I include specific comments below.

I plan to look this over during the next round of revisions, once the authors have provided more detail on the methods.

- [ x] **Examples:** (that run successfully locally) for all exported functions.
Great job here!

#### Functionality

- [ x] **Installation:** Installation succeeds as documented.
- [x ] **Functionality:** Any functional claims of the software have been confirmed.
- [ x] **Performance:** Any performance claims of the software have been confirmed.

Will check this next round.

- [ x] Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer (“rev” role) in the package DESCRIPTION file.

### Review Comments

#### Overall
- For this round of comments I focused mostly on the vignette. I was able to work through all of the examples without any issue. However, there is one component that is missing from this. I think it will be important to add a section or separate page that focuses on how to run this using your own dataset. For example, if folks start with Raven selection tables, sounds files, and metadata of distance and other information, how do they get started using these methods? The examples in the vignette work great, but I think some work could be done to help people get their own data in the pipeline with this.

#### README
-   It would be helpful to include an example or type of research question. This may be one of the first exposure people have to this type of work, so a few examples would be helpful
-   What does the check part mean? A figure legend describing the diagram would be helpful.

#### Synthesize and align test sounds
- I was quite impressed with how smoothly I was able to work through this section of the vignette- nice work!

- You might want to include a warning about playing the sound synthetic master file (!). It gave me quite a shock!

- I can imagine there will be times when spectrogram cross-correlation won't work to time align files. Perhaps adding something about using caution and the need to manually check would be good.

#### Quantify degradation
- I realize this is an R package vignette and not a peer-reviewed paper, but I think including some pros/cons of different measures would be helpful. How should someone who is coming to this for the first time decide which values to measure and report?

- For all of the different types of degradation measures it would be good to include what units they are in (e.g. decibels, etc), as this will facilitate folks who want to use this for publications and to report their results using best practices.

-   “However, if there is another test sound from the same ‘sound.id’ at a shorter distance in other transects, it will be used as reference instead.” Should this be the default? What if transects are very different?

-   “The function also checks that the information ‘X’ is in the right format so it wont produce errors in downstream analysis (see ‘X’ argument description for details on format).” I am not sure what the X argument is?

-   Typo? Right above ‘blur ratio’: “Before This is one the plots generated using the example data”.

-   “The env.smooth argument could change envelope shapes and related measurements. The following code sets env.smooth = 800:” Could you provide more information on this, why would you want to change it, and what would the expected changes be?

-   Can you include a citation for ‘spectrum blur ratio’?

-   For both blur ratios, it would be good to briefly explain how you calculate the mismatch.

- SNR can be measured many different ways (e.g. from waveform or spectrogram). I think there need to be a few lines added here explaining which approach is used to measure SNR.

- Some of the SNR values in the example seem strange, such as -37? More detail on how it is calculated and what is actually reported will be helpful here.

- What should the range of TSR be? All the examples are negative.

- The references info is missing from this page
