+++ R/internal_functions.R
@@ -1218,8 +1218,8 @@
   return(envcor)
 }
 
-# function to extract mean envelopes
-mean.env <- function(y, wl, ovlp, X, path, bp) {
+# function to extract mean envelopes or RMS of envelopes
+.mean.env <- function(y, wl, ovlp, X, path, bp, rms = FALSE) {
   # read sound clip
   clp <-
     warbleR::read_sound_file(
@@ -1251,8 +1251,11 @@ mean.env <- function(y, wl, ovlp, X, path, bp) {
       )
   }
   
-  sig_env <-
-    mean(warbleR::envelope(x = clp@left))
+  
+  sig_env <-   if (!rms)
+    mean(warbleR::envelope(x = clp@left)) else
+      seewave::rms(warbleR::envelope(x = clp@left))
+  
   
   return(sig_env)
 }
@@ -1294,7 +1297,7 @@ mean.env <- function(y, wl, ovlp, X, path, bp) {
   return(cor.spctr)
 }
 
-.exc_att <- function(y, X, tp, gn) {
+.exc_att <- function(y, X) {
   # get names of sound and reference
   sgnl <- X$.sgnl.temp[y]
   rfrnc <- X$reference[y]
@@ -1308,20 +1311,10 @@ mean.env <- function(y, wl, ovlp, X, path, bp) {
     dist_REF <- X$distance[X$.sgnl.temp == rfrnc]
     dist_SIG <- X$distance[y]
     
-    ks <- X$sig_env[y] / sig_env_REF
-    
-    # type Dabelsteen
-    if (tp == "Dabelsteen") {
-      ea <-
-        (-20 * log(ks)) - (6 / (2 * (dist_SIG - dist_REF))) + gn
-    }
-    
-    if (tp == "Darden") {
-      # EA = g - 20 log(d / 10) - 20 log(k)
-      ea <-
-        gn - 20 * log10(dist_SIG / 10) - 20 * log(ks)
-    }
+      # excess attenuation = (total attenuation - spheric spreading attenuation) 
+      ea <- (-20 * log10(sig_env_REF / X$sig_env[y])) - (20 * log10(1 / dist_SIG))
   }
+  
   return(ea)
 }
 
@@ -1785,7 +1778,7 @@ mean.env <- function(y, wl, ovlp, X, path, bp) {
   
   # adjust wl based on hop.size
   if (is.null(wl)) {
-    wl <- if (is_extended_selection_table(X))
+    wl <- if (warbleR::is_extended_selection_table(X))
       round(attr(X, "check.results")$sample.rate[1] * hop.size, 0) else
       round(
         read_sound_file(
@@ -3305,7 +3298,7 @@ mean.env <- function(y, wl, ovlp, X, path, bp) {
       X$end[X$sound.files == u] + step_sum_vector[u]
     
     # fix internally for extended selection tables
-    if (is_selection_table(X) | is_extended_selection_table(X)) {
+    if (is_selection_table(X) | warbleR::is_extended_selection_table(X)) {
       attributes(X)$check.res$start[attributes(X)$check.res$sound.files == u] <-
         attributes(X)$check.res$start[X$sound.files == u] + step_sum_vector[u]
       attributes(X)$check.res$end[attributes(X)$check.res$sound.files == u] <-
@@ -3319,6 +3312,110 @@ mean.env <- function(y, wl, ovlp, X, path, bp) {
   
   return(X)
 }
+
+# make markers for master sound files used by master_sound_file()
+.make_markers <- function(X, flim, sampling_rate, cex){
+  
+  # remove plots at the end
+  on.exit(try(dev.off(), silent = TRUE))
+  
+  # check if ghost script is installed
[…]
