+++ R/master_sound_file.R
@@ -268,7 +268,7 @@ master_sound_file <-
     # add start & end markers
     sel.tab$sound.id <- if (!is.null(X$sound.id)) {
       c("start_marker", X$sound.id, "end_marker")
-    } else if (is_extended_selection_table(X)) {
+    } else if (warbleR::is_extended_selection_table(X)) {
       c("start_marker", X$sound.files, "end_marker")
     } else {
       c("start_marker",
@@ -307,149 +307,3 @@ master_sound_file <-
     
     return(sel.tab)
   }
-
-
-# make markers for master sound files used by master_sound_file()
-.make_markers <- function(X, flim, sampling_rate, cex){
-  
-  # remove plots at the end
-  on.exit(try(dev.off(), silent = TRUE))
-  
-  # check if ghost script is installed
-  gsexe <- tools::find_gs_cmd()
-  
-  # warning if ghostscript not found
-  if (!nzchar(gsexe)) {
-    .message(paste(
-      cli::cli_text(
-        .colortext(
-          "warning: GhostScript was not found. It produces clearer start and end marker. You can download it from {.url https://ghostscript.com}. ",
-          as = "red"
-        )
-      ),
-      .colortext("Using 'png()' instead.", as = "red")
-    ))
-  }
-  
-  # set frequency range for markers
-  if (is.null(flim)) {
-    flim <- c(0, attr(X, "check.results")$sample.rate[1] / 2)
-  }
-  
-  # save par settings
-  oldpar <- par(no.readonly = TRUE)
-  on.exit(par(
-    mar =  oldpar$mar,
-    mfcol =  oldpar$mfcol,
-    mfrow =  oldpar$mfrow
-  ))
-  
-  # save image of start marker in temporary directory
-  if (!nzchar(gsexe)) {
-    grDevices::png(filename = file.path(tempdir(), "strt_mrkr-img.png"),
-                   pointsize = 10)
-  }
-  
-  # remove margins in graphic device
-  par(mar = rep(0, 4),
-      mfrow = c(1, 1),
-      mfcol = c(1, 1))
-  
-  # empty plot
-  plot(
-    0,
-    type = "n",
-    axes = FALSE,
-    ann = FALSE,
-    xlim = c(0, 1),
-    ylim = c(0, 1)
-  )
-  
-  # add text
-  text(
-    x = 0.5,
-    y = 0.5,
-    labels = "*start+",
-    cex = cex,
-    font = 2
-  )
-  
-  # save image of start marker in temporary directory
-  if (nzchar(gsexe)) {
-    grDevices::dev2bitmap(file.path(tempdir(), "strt_mrkr-img.png"),
-                          type = "pngmono",
-                          res = 30)
-  }
-  
-  # close graph
-  dev.off()
-  
-  # make image of start marker
-  strt_mrkr <-
-    warbleR::image_to_wave(
-      file = file.path(tempdir(), "strt_mrkr-img.png"),
-      plot = FALSE,
-      flim = flim,
-      samp.rate = sampling_rate / 1000
-    )
-  
-  # remove image file
-  unlink(file.path(tempdir(), "strt_mrkr-img.png"))
-  
-  # save image of end marker in temporary directory
-  if (!nzchar(gsexe)) {
-    grDevices::png(filename = file.path(tempdir(), "end_mrkr-img.png"),
-                   pointsize = 10)
-  }
-  
-  # remove margins in graphic device
-  par(mar = rep(0, 4),
-      mfrow = c(1, 1),
-      mfcol = c(1, 1))
-  
-  # empty plot
-  plot(
-    0,
-    type = "n",
[…]
