+++ vignettes/align_test_sounds.Rmd
@@ -110,7 +110,7 @@ library(baRulho)
 library(Rraven)
 
 # synthesize
-synth_est <- synth_sounds(
+synth_est <- baRulho::synth_sounds(
   mar = 0.1,
   frequencies = c(0.5, 1:5),
   durations = 0.1,
@@ -148,7 +148,7 @@ The function can add complexity to additional acoustic dimensions. For instance 
 ```{r, eval = TRUE, echo = TRUE, fig.height=5, dpi=120, message=FALSE, warning=FALSE}
 
 # synthesize
-synth_est2 <- synth_sounds(
+synth_est2 <- baRulho::synth_sounds(
   mar = 0.01,
   sig2 = 0.5,
   frequencies = 2:4,
@@ -187,7 +187,6 @@ seewave::spectro(
 
 Note that the output is an extended selection table that contains the time-frequency annotations for all synthetic sounds:
 
-
 ```{r, eval = TRUE}
 # check first 6 rows
 head(as.data.frame(synth_est2))
@@ -214,7 +213,7 @@ The following example shows how to create a master sound file using the syntheti
 ```{r master sound file, eval = TRUE, echo = TRUE, fig.show='hide'}
 
 # create master sound file
-synth_master_annotations <- master_sound_file(
+synth_master_annotations <- baRulho::master_sound_file(
   X = synth_est,
   file.name = "synthetic_master",
   dest.path = tempdir(),
@@ -274,8 +273,11 @@ est <- warbleR::selection_table(
   pb = FALSE
 )
 
+# add sound_id column
+est$sound.id <- paste0(abbreviate(est$sound.files), est$selec)
+
 # create master sound file
-master_annotations <- master_sound_file(
+master_annotations <- baRulho::master_sound_file(
   X = est,
   file.name = "example_master",
   dest.path = tempdir(),
@@ -422,11 +424,12 @@ markers_position <- baRulho::find_markers(X = master_annotations, pb = FALSE)
 &nbsp;
 
 The position of the markers is determined as the highest spectrogram cross-correlation value for each marker using the functions `template_correlator()` and `template_detector` from the package [ohun](https://CRAN.R-project.org/package=ohun). If two or more markers are used the function computes an additional column, 'time.mismatch', that compares the time difference between the two markers in the test-files against that in the master sound file. In a perfect detection the value must be 0. So this number can be used as a measure of error. In some cases precision can improve by increasing the time resolution of the spectrogram, that is, by lowering the size of the time window. This is done with the argument 'hop.size'. By default it is 11.6 ms (which is equivalent to a 512 samples window length in a 44.1 kHz file). Let's try 4 ms: 
+
 ```{r, eval = FALSE}
 
 # lower window length
 markers_position <-
-  find_markers(X = master_annotations, hop.size = 4)
+  baRulho::find_markers(X = master_annotations, hop.size = 4)
 
 markers_position
 ```
@@ -458,17 +461,6 @@ aligned_tests
 ```{r, eval = TRUE, echo = FALSE}
 
 print(TRUE)
-```
-
-```{r, eval = FALSE, echo = FALSE, results='asis'}
-
-kbl <- kable(aligned_tests, align = "c", row.names = F, format = "html", escape = F)
-
-kbl <- kableExtra::kable_styling(kbl, bootstrap_options = "striped", font_size = 14)
-
-kbl <- kableExtra::scroll_box(kbl, width = "800px", height = "300px")
-
-kbl
 ```
 &nbsp;
 
