+++ vignettes/align_test_sounds.Rmd
@@ -391,47 +391,28 @@ master_annotations <- as.data.frame(master_est)
 
 To find the location of the start marker on the test sound files we use the functions `find_markers()` to run a cross-correlation of one or more markers across the test files in order to determine the exact time in which each marker is found: 
 
-```{r, eval = FALSE}
-options(sound.files.path = tempdir())
+```{r, eval = TRUE}
 
-markers_position <- baRulho::find_markers(X = master_annotations)
+markers_position <-
+  baRulho::find_markers(X = master_annotations, path = tempdir())
 
 markers_position
-```
-
-```{r, eval = TRUE, echo=FALSE, results='asis'}
-
-# set path globally (so no need to write it each time)
-options(sound.files.path = tempdir())
-
-markers_position <- baRulho::find_markers(X = master_annotations, pb = FALSE)
-
-# kbl <-
-#   kable(
-#     markers_position,
-#     align = "c",
-#     row.names = F,
-#     format = "html",
-#     escape = F
-#   )
-# 
-# kbl <-
-#   kableExtra::kable_styling(kbl, bootstrap_options = "striped", font_size = 14)
-# 
-# kbl
 ```
 
 &nbsp;
 
 The position of the markers is determined as the highest spectrogram cross-correlation value for each marker using the functions `template_correlator()` and `template_detector` from the package [ohun](https://CRAN.R-project.org/package=ohun). If two or more markers are used the function computes an additional column, 'time.mismatch', that compares the time difference between the two markers in the test-files against that in the master sound file. In a perfect detection the value must be 0. So this number can be used as a measure of error. In some cases precision can improve by increasing the time resolution of the spectrogram, that is, by lowering the size of the time window. This is done with the argument 'hop.size'. By default it is 11.6 ms (which is equivalent to a 512 samples window length in a 44.1 kHz file). Let's try 4 ms: 
 
-```{r, eval = FALSE}
+```{r, eval = TRUE}
 
 # lower window length
 markers_position <-
-  baRulho::find_markers(X = master_annotations, hop.size = 4)
+  baRulho::find_markers(X = master_annotations,
+                        hop.size = 4,
+                        path = tempdir())
 
 markers_position
+
 ```
 &nbsp;
 
@@ -444,7 +425,8 @@ aligned_tests <-
     X = master_annotations,
     Y = markers_position,
     by.song = TRUE,
-    remove.markers = FALSE
+    remove.markers = FALSE,
+    path = tempdir()
   )
 ```
 
@@ -472,6 +454,7 @@ We can also check the precision of the alignment by looking at the spectrograms.
 
 baRulho::plot_aligned_sounds(
   X = aligned_tests,
+  path = tempdir(),
   dest.path = tempdir(),
   duration = 2.4,
   ovlp = 0
@@ -492,7 +475,12 @@ In some cases alignments might not be totally accurate. The function `manual_rea
 
 ```{r, eval = FALSE}
 
-baRulho::manual_realign(X = aligned_tests, Y = master_annotations, duration = 2.4)
+baRulho::manual_realign(
+  X = aligned_tests,
+  Y = master_annotations,
+  duration = 2.4,
+  path = tempdir()
+)
 
 ```
 
@@ -536,7 +524,7 @@ Now we can re-aligned test sounds:
 
 # re align
 rts <-
-  baRulho::auto_realign(X = unaligned_test_sounds_est, pb = FALSE)
+  baRulho::auto_realign(X = unaligned_test_sounds_est)
 
 ```
 
