### 2024-01-06 598b18e
remove options() calls
