### 2024-01-06 3a8c41a
env.smooth = 800 added
