### 2024-01-06 7e22e55
use scale_color_viridis always
