+++ R/manual_realign.R
@@ -40,9 +40,6 @@
 #' @seealso \code{\link{auto_realign}}; \code{\link{find_markers}}; \code{\link{align_test_files}}
 #' @examples
 #' {
-#'   # set temporary directory
-#'   td <- tempdir()
-#'
 #'   # load example data
 #'   data("master_est")
 #'
@@ -51,30 +48,28 @@
 #'   # This doesn't have to be done with your own data as you will
 #'   # have them as sound files already.
 #'   for (i in unique(test_sounds_est$sound.files)[1:2]) {
-#'     writeWave(object = attr(test_sounds_est, "wave.objects")[[i]], file.path(td, i))
+#'   writeWave(object = attr(test_sounds_est, "wave.objects")[[i]], file.path(tempdir(), i))
 #'   }
 #'
 #'   # save master file
-#'   writeWave(object = attr(master_est, "wave.objects")[[1]], file.path(td, "master.wav"))
-#'
-#'   # set path and no progress bar in global options
-#'   options(sound.files.path = td, pb = FALSE)
+#'   writeWave(object = attr(master_est, "wave.objects")[[1]], file.path(tempdir(), "master.wav"))
 #'
 #'   # get marker position
-#'   markers <- find_markers(X = master_est, test.files = unique(test_sounds_est$sound.files)[2])
+#'   markers <- find_markers(X = master_est, test.files = unique(test_sounds_est$sound.files)[2], 
+#'   path = tempdir())
 #'
 #'   # align all test sounds
-#'   alg.tests <- align_test_files(X = master_est, Y = markers, pb = FALSE)
+#'   alg.tests <- align_test_files(X = master_est, Y = markers)
 #'
 #'   # add error to alignment
 #'   lag <- (as.numeric(as.factor(alg.tests$sound.files)) - 2) / 30
 #'   alg.tests$start <- alg.tests$start + lag
 #'   alg.tests$end <- alg.tests$end + lag
 #'
-#'   \dontrun{
-#'   realigned_est <- manual_realign(X = alg.tests, Y = master_df, duration = 2,
+#'   if(interactive()){
+#'   realigned_est <- manual_realign(X = alg.tests, Y = master_est, duration = 2,
 #'   ovlp = 50, hop.size = 14, collevels = seq(-140, 0, 5), palette = mako,
-#'   ext.window = F)
+#'   ext.window = FALSE)
 #'  }
 #' }
 #' @author Marcelo Araya-Salas (\email{marcelo.araya@@ucr.ac.cr})
