+++ man/manual_realign.Rd
@@ -70,9 +70,6 @@ This function allows the interactive adjustment of the alignment of test sound f
 }
 \examples{
 {
-  # set temporary directory
-  td <- tempdir()
-
   # load example data
   data("master_est")
 
@@ -81,31 +78,24 @@ This function allows the interactive adjustment of the alignment of test sound f
   # This doesn't have to be done with your own data as you will
   # have them as sound files already.
   for (i in unique(test_sounds_est$sound.files)[1:2]) {
-    writeWave(object = attr(test_sounds_est, "wave.objects")[[i]], file.path(td, i))
+  writeWave(object = attr(test_sounds_est, "wave.objects")[[i]], file.path(tempdir(), i))
   }
 
   # save master file
-  writeWave(object = attr(master_est, "wave.objects")[[1]], file.path(td, "master.wav"))
-
-  # set path and no progress bar in global options
-  options(sound.files.path = td, pb = FALSE)
+  writeWave(object = attr(master_est, "wave.objects")[[1]], file.path(tempdir(), "master.wav"))
 
   # get marker position
-  markers <- find_markers(X = master_est, test.files = unique(test_sounds_est$sound.files)[2])
+  markers <- find_markers(X = master_est, test.files = unique(test_sounds_est$sound.files)[2], 
+  path = tempdir())
 
   # align all test sounds
-  alg.tests <- align_test_files(X = master_est, Y = markers, pb = FALSE)
+  alg.tests <- align_test_files(X = master_est, Y = markers)
 
   # add error to alignment
   lag <- (as.numeric(as.factor(alg.tests$sound.files)) - 2) / 30
   alg.tests$start <- alg.tests$start + lag
   alg.tests$end <- alg.tests$end + lag
 
-  \dontrun{
-  realigned_est <- manual_realign(X = alg.tests, Y = master_df, duration = 2,
-  ovlp = 50, hop.size = 14, collevels = seq(-140, 0, 5), palette = mako,
-  ext.window = F)
- }
 }
 }
 \references{
