### 2024-01-07 e4de67b
remove dontrun blur_ratio()
