### 2024-01-07 2646d69
added  if (requireNamespace("ggplot2", quietly = TRUE))
