+++ man/spectrum_blur_ratio.Rd
@@ -72,7 +72,7 @@ X <- set_reference_sounds(X = test_sounds_est)
   # make distance a factor for plotting
   spctr$distance <- as.factor(spctr$distance)
 
-  \dontrun{
+  if (requireNamespace("ggplot2", quietly = TRUE)){
   # plot
   ggplot(spctr[spctr$freq > 0.3, ], aes(y = amp, x = freq,
   col = distance)) +
