+++ R/blur_ratio.R
@@ -1,13 +1,6 @@
 #' Measure blur ratio in the time domain
 #'
 #' \code{blur_ratio} measures blur ratio in sounds referenced in an extended selection table.
-#' @usage blur_ratio(X, parallel = NULL, cores = getOption("mc.cores", 1),
-#' pb = getOption("pb", TRUE),
-#' env.smooth = getOption("env.smooth", 200),
-#' output = NULL, envelopes = FALSE,
-#' hop.size = getOption("hop.size", 11.6), wl = getOption("wl", NULL),
-#' ovlp = getOption("ovlp", 70), n.samples = 100,
-#' path = getOption("sound.files.path", "."))
 #' @param X The output of \code{\link{set_reference_sounds}} which is an object of class 'data.frame', 'selection_table' or 'extended_selection_table' (the last 2 classes are created by the function \code{\link[warbleR]{selection_table}} from the warbleR package) with the reference to the test sounds . Must contain the following columns: 1) "sound.files": name of the .wav files, 2) "selec": unique selection identifier (within a sound file), 3) "start": start time and 4) "end": end time of selections, 5)  "bottom.freq": low frequency for bandpass, 6) "top.freq": high frequency for bandpass, 7) "sound.id": ID of sounds used to identify counterparts across distances and 8) "reference": identity of sounds to be used as reference for each test sound (row). See \code{\link{set_reference_sounds}} for more details on the structure of 'X'.
 #' @param parallel DEPRECATED. Use 'cores' instead.
 #' @param cores Numeric vector of length 1. Controls whether parallel computing is applied by specifying the number of cores to be used. Default is 1 (i.e. no parallel computing).
@@ -49,6 +42,8 @@
 #'
 #'   if (requireNamespace("ggplot2", quietly = TRUE)){
 #'   # plot
+#'   library(ggplot2)
+#'   
 #'   ggplot(envs, aes(x= time, y = amp, col = distance)) +
 #'   geom_line() + facet_wrap(~ sound.id) +
 #'   scale_color_viridis_d() +
