+++ man/attenuation.Rd
@@ -33,7 +33,7 @@ Calculate the geometric, atmospheric and habitat attenuation and the overall exp
 \examples{
 {
   # measure attenuation
-  attenuation(f = 2000, dist = 50, dist0 = 1)
+  attenuation(frequency = 2000, dist = 50, dist0 = 1)
 }
 }
 \references{
