+++ man/blur_ratio.Rd
@@ -4,13 +4,20 @@
 \alias{blur_ratio}
 \title{Measure blur ratio in the time domain}
 \usage{
-blur_ratio(X, parallel = NULL, cores = getOption("mc.cores", 1),
-pb = getOption("pb", TRUE),
-env.smooth = getOption("env.smooth", 200),
-output = NULL, envelopes = FALSE,
-hop.size = getOption("hop.size", 11.6), wl = getOption("wl", NULL),
-ovlp = getOption("ovlp", 70), n.samples = 100,
-path = getOption("sound.files.path", "."))
+blur_ratio(
+  X,
+  parallel = NULL,
+  cores = getOption("mc.cores", 1),
+  pb = getOption("pb", TRUE),
+  env.smooth = getOption("env.smooth", 200),
+  output = NULL,
+  envelopes = FALSE,
+  hop.size = getOption("hop.size", 11.6),
+  wl = getOption("wl", NULL),
+  ovlp = getOption("ovlp", 70),
+  n.samples = 100,
+  path = getOption("sound.files.path", ".")
+)
 }
 \arguments{
 \item{X}{The output of \code{\link{set_reference_sounds}} which is an object of class 'data.frame', 'selection_table' or 'extended_selection_table' (the last 2 classes are created by the function \code{\link[warbleR]{selection_table}} from the warbleR package) with the reference to the test sounds . Must contain the following columns: 1) "sound.files": name of the .wav files, 2) "selec": unique selection identifier (within a sound file), 3) "start": start time and 4) "end": end time of selections, 5)  "bottom.freq": low frequency for bandpass, 6) "top.freq": high frequency for bandpass, 7) "sound.id": ID of sounds used to identify counterparts across distances and 8) "reference": identity of sounds to be used as reference for each test sound (row). See \code{\link{set_reference_sounds}} for more details on the structure of 'X'.}
@@ -54,12 +61,12 @@ Blur ratio measures the degradation of sound as a change in sound power in the t
   data("test_sounds_est")
 
  # add reference to X
- X <- set_reference_sounds(X = test_sounds_est)
+  X <- set_reference_sounds(X = test_sounds_est)
   blur_ratio(X = X)
 
   # using method 2
 X <- set_reference_sounds(X = test_sounds_est, method = 2)
-  # blur_ratio(X = X)
+  blur_ratio(X = X)
 
   # get envelopes
   br <- blur_ratio(X = X, envelopes = TRUE)
