+++ R/auto_realign.R
@@ -1,9 +1,6 @@
 #' Fix misaligned start and end of test sounds
 #'
 #' \code{auto_realign} fixes misaligned start and end of test sounds in an extended selection table using spectrographic cross-correlation
-#' @usage auto_realign(X, parallel = NULL, cores = getOption("mc.cores", 1),
-#' pb = getOption("pb", TRUE), hop.size = getOption("hop.size", 11.6),
-#' wl = getOption("wl", NULL), ovlp = getOption("ovlp", 90), wn = 'hanning')
 #' @param X object of class 'extended_selection_table' created by the function \code{\link[warbleR]{selection_table}} from the warbleR package. The object must include the following additional columns: 'sound.id', 'bottom.freq' and 'top.freq'.
 #' @param parallel DEPRECATED. Use 'cores' instead.
 #' @param cores Numeric vector of length 1. Controls whether parallel computing is applied by specifying the number of cores to be used. Default is 1 (i.e. no parallel computing).
