+++ R/align_test_files.R
@@ -1,9 +1,6 @@
 #' Align test sound files
 #'
 #' \code{align_test_files} aligns test (re-recorded) sound files.
-#' @usage align_test_files(X, Y, output = NULL, path = getOption("sound.files.path", "."),
-#' by.song = TRUE, marker = NULL, cores = getOption("mc.cores", 1),
-#' pb = getOption("pb", TRUE), ...)
 #' @param X object of class 'data.frame', 'selection_table' or 'extended_selection_table' (the last 2 classes are created by the function \code{\link[warbleR]{selection_table}} from the warbleR package). This should be the same data than that was used for finding the position of markers in \code{\link{find_markers}}. It should also contain a 'sound.id' column that will be used to label re-recorded sounds according to their counterpart in the master sound file.
 #' @param Y object of class 'data.frame' with the output of \code{\link{find_markers}}. This object contains the position of markers in the re-recorded sound files. If more than one marker is supplied for a sound file only the one with the highest correlation score ('scores' column in 'X') is used.
 #' @param output DEPRECATED. Now the output format mirrors the class of the input 'X'.
