### 2024-01-07 3610d04
comments to magic numbers
