+++ R/auto_realign.R
@@ -1,38 +1,38 @@
-#' Fix misaligned start and end of test sounds
+#' Fix small misalignments in the time position test sounds
 #'
-#' \code{auto_realign} fixes misaligned start and end of test sounds in an extended selection table using spectrographic cross-correlation
+#' \code{auto_realign} fixes small misalignments in the time position of test sounds in an extended selection table using spectrographic cross-correlation
 #' @inheritParams template_params
 #' @param X object of class 'extended_selection_table' created by the function \code{\link[warbleR]{selection_table}} from the warbleR package. The object must include the following additional columns: 'sound.id', 'bottom.freq' and 'top.freq'.
+#' @param Y object of class 'extended_selection_table' (a class created by the function \code{\link[warbleR]{selection_table}} from the warbleR package) with the master sound file annotations. This should be the same data than that was used for finding the position of markers in \code{\link{find_markers}}. It should also contain a 'sound.id' column.
 #' @param ovlp Numeric vector of length 1 specifying the percentage of overlap between two
-#' consecutive windows, as in \code{\link[seewave]{spectro}}. Default is 90. High values of ovlp
-#' slow down the function but produce more accurate results.
-#' @param wn A character vector of length 1 specifying the window name as in \code{\link[seewave]{ftwindow}}.
+#' consecutive windows, as in \code{\link[seewave]{spectro}}. Default is 90. High values slow down the function but produce more accurate results.
 #' @return Object 'X' in which time parameters (columns 'start' and 'end') have been tailored to more closely match the start and end of the reference sound.
 #' @export
 #' @name auto_realign
-#' @details This function uses spectrographic cross-correlation to align the position in time of sounds with regard to a reference sound. The sound recorded at the closest distance to the source is used as reference. Precise alignment is crucial for downstream measures of sound degradation. The function calls warbleR's \code{\link[warbleR]{cross_correlation}} internally to align sounds using cross-correlation. The output extended selection table contains the new start and end values after alignment.
+#' @details Precise alignment is crucial for downstream measures of sound degradation. This function uses spectrographic cross-correlation to align the position in time of test sounds. The master sound file is used as reference. The function calls warbleR's \code{\link[warbleR]{cross_correlation}} internally to align sounds using cross-correlation. The output extended selection table contains the new start and end values after alignment. Note that this function only works to further improve alignments if the estimated position of the test sound is already close to the actual position. Note that both 'X' and 'Y' must be extended selection tables sensu \code{\link[warbleR]{selection_table}}.
+#' 
 #' @examples {
 #'   # load example data
 #'   data("test_sounds_est")
-#'
+#'   data("master_est")
+#'   
 #'   # create "unaligned_test_sounds_est" by
-#'   # adding noise to "test_sounds_est" start and end
+#'   # adding error to "test_sounds_est" start and end
 #'   unaligned_test_sounds_est <- test_sounds_est
 #'   set.seed(123)
[…]
