+++ R/signal_to_noise_ratio.R
@@ -54,14 +54,12 @@
 signal_to_noise_ratio <-
   function(X,
            mar = NULL,
-           parallel = NULL,
            cores = getOption("mc.cores", 1),
            pb = getOption("pb", TRUE),
            eq.dur = FALSE,
            noise.ref = "adjacent",
            type = 1,
            bp = "freq.range",
-           output = NULL,
            hop.size = getOption("hop.size", 1),
            wl = getOption("wl", NULL),
            ovlp = getOption("ovlp", 0),
