### 2024-01-09 8eadbdf
Update CITATION.cff
