### 2024-01-09 2be41e7
added rlang::check_installed
