+++ R/spectrum_blur_ratio.R
@@ -35,8 +35,9 @@
 #'   # make distance a factor for plotting
 #'   spctr$distance <- as.factor(spctr$distance)
 #'
-#'   if (requireNamespace("ggplot2", quietly = TRUE)){
+#'   
 #'   # plot
+#'   rlang::check_installed("ggplot2"))
 #'   library(ggplot2)
 #'   
 #'   ggplot(spctr[spctr$freq > 0.3, ], aes(y = amp, x = freq,
@@ -47,7 +48,6 @@
 #'   labs(x = "Frequency (kHz)", y = "Amplitude (PMF)") +
 #'   coord_flip() +
 #'   theme_classic()
-#'   }
 #' }
 #'
 #' @seealso \code{\link{blur_ratio}}
