+++ R/blur_ratio.R
@@ -32,8 +32,9 @@
 #'   # make distance a factor for plotting
 #'   envs$distance <- as.factor(envs$distance)
 #'
-#'   if (requireNamespace("ggplot2", quietly = TRUE)){
+#'   
 #'   # plot
+#'   rlang::check_installed("ggplot2")
 #'   library(ggplot2)
 #'   
 #'   ggplot(envs, aes(x= time, y = amp, col = distance)) +
@@ -41,7 +42,7 @@
 #'   scale_color_viridis_d() +
 #'   labs(x = "Time (s)", y = "Amplitude (PMF)") +
 #'   theme_classic()
-#'   }
+#'   
 #' }
 #'
 #' @author Marcelo Araya-Salas (\email{marcelo.araya@@ucr.ac.cr})
