+++ man/spectrum_blur_ratio.Rd
@@ -74,8 +74,9 @@ X <- set_reference_sounds(X = test_sounds_est)
   # make distance a factor for plotting
   spctr$distance <- as.factor(spctr$distance)
 
-  if (requireNamespace("ggplot2", quietly = TRUE)){
+  
   # plot
+  rlang::check_installed("ggplot2"))
   library(ggplot2)
   
   ggplot(spctr[spctr$freq > 0.3, ], aes(y = amp, x = freq,
@@ -86,7 +87,6 @@ X <- set_reference_sounds(X = test_sounds_est)
   labs(x = "Frequency (kHz)", y = "Amplitude (PMF)") +
   coord_flip() +
   theme_classic()
-  }
 }
 
 }
