+++ NAMESPACE
@@ -55,6 +55,7 @@ importFrom(methods,is)
 importFrom(ohun,template_correlator)
 importFrom(ohun,template_detector)
 importFrom(rlang,arg_match)
+importFrom(rlang,check_installed)
 importFrom(stats,aggregate)
 importFrom(stats,approx)
 importFrom(stats,ave)
