### 2024-01-09 #609 @DenaJGibbon
It was 4-5 for me. Thanks!

On Tue, Jan 9, 2024 at 7:40 AM ropensci-review-bot ***@***.***>
wrote:

> Logged review for *mikemahoney218* (hours: 6)
>
