+++ R/detection_distance.R
@@ -53,8 +53,12 @@ detection_distance <-
            max.distance = 1000,
            resolution = 0.1,
            subtract.bgn = TRUE,
-           envelope = "abs",
+           envelope = c("abs", "hil"),
            mar = NULL) {
+    
+    # assign a value to type
+    envelope <- rlang::arg_match(envelope)
+             
     # check arguments
     arguments <- as.list(base::match.call())
     
