### 2024-01-09 3bf53d8
return file path when plotting images
