### 2024-01-09 c18f56a
fix color param description
