### 2024-01-09 9c39807
rename type argument in tsr and snr
