+++ README.Rmd
@@ -18,7 +18,11 @@ editor_options:
 
 [baRulho](https://cran.r-project.org/package=baRulho) is intended to facilitate the implementation of (animal) sound transmission experiments, which typically aim to quantify changes in signal structure when transmitted in a given habitat by broadcasting and re-recording animal sounds at increasing distances. A common sequence of steps to experimentally test hypotheses related to sound transmission is depicted in the following diagram:
 
- <center><img src="man/figures/analysis_workflow.jpg" alt="analysis workflow" width="80%"></center>
+```{r, echo=FALSE, out.width="100%", fig.align='center', fig.cap= "Figure 1. Diagram depicting a typical workflow of experiments working on signal tranmission and degradation. Nodes with black font indicate steps that can be conducted using baRulho functions. Blue nodes denote the functions that can be used at those steps."}
+
+knitr::include_graphics("man/figures/analysis_workflow.png")
+
+```
 
 &nbsp;
 
