### 2024-01-12 d244719
add warning about bad alignment
