+++ vignettes/align_test_sounds.Rmd
@@ -411,7 +411,7 @@ markers_position
 
 &nbsp;
 
-The position of the markers is determined as the highest spectrogram cross-correlation value for each marker using the functions `template_correlator()` and `template_detector` from the package [ohun](https://CRAN.R-project.org/package=ohun). If two or more markers are used the function computes an additional column, 'time.mismatch', that compares the time difference between the two markers in the test-files against that in the master sound file. In a perfect detection the value must be 0. So this number can be used as a measure of error. In some cases precision can improve by increasing the time resolution of the spectrogram, that is, by lowering the size of the time window. This is done with the argument 'hop.size'. By default it is 11.6 ms (which is equivalent to a 512 samples window length in a 44.1 kHz file). Let's try 4 ms: 
+The position of the markers is determined as the highest spectrogram cross-correlation value for each marker using the functions `find_markers()`. If two or more markers are used the function computes an additional column, 'time.mismatch', that compares the time difference between the two markers in the test files against that in the master sound file. In a perfect detection the value must be 0. So this number can be used as a measure of error. In some cases precision can improve by increasing the time resolution of the spectrogram, that is, by lowering the size of the time window. This is done with the argument 'hop.size'. By default it is 11.6 ms (which is equivalent to a 512 samples window length in a 44.1 kHz file). Let's try 4 ms: 
 
 ```{r, eval = TRUE}
 
@@ -458,30 +458,37 @@ print(TRUE)
 
 This format is similar to data frames, although it also contains all the acoustic data as attributes so there is no need to refer to sound files in a directory any more.  [baRulho](https://marce10.github.io/baRulho/) functions that quantify degradation can take data frames, selection tables and extended selection tables.  
 
-We can also check the precision of the alignment by looking at the spectrograms. The function `plot_aligned_sounds()` can create spectrograms to visually inspect alignment precision on test sound files. The spectrograms are saved as individual JPEG image files (in 'dest.path' if supplied). One file is created for each sound file:
+We can check the precision of the alignment by looking at the spectrograms. The function `plot_aligned_sounds()` can create spectrograms to visually inspect alignment precision on test sound files. The spectrograms are saved as individual JPEG image files (in 'dest.path' if supplied). One file is created for each sound file:
 
-```{r, eval = FALSE}
+```{r, eval = TRUE}
 
-baRulho::plot_aligned_sounds(
+aligned_imgs <- baRulho::plot_aligned_sounds(
   X = aligned_tests,
   path = tempdir(),
   dest.path = tempdir(),
   duration = 2.4,
   ovlp = 0
 )
+
+aligned_imgs
+
 ```
 
  These are two of the files generated by the code above:
 
- <img src="plot_align_10m_open.jpeg" alt="test 1">
+```{r, echo=FALSE, out.width="100%"}
 
- <img src="plot_align_30m_closed.jpeg" alt="test 2">
+knitr::include_graphics(aligned_imgs[[1]])
+
+knitr::include_graphics(aligned_imgs[[4]])
+
+```
 
 &nbsp;
 
 ## Manually fixing alignment
 
-In some cases alignments might not be totally accurate. The function `manual_realign()` allows users to interactively adjust the alignment of test sound files produced by `align_test_files()`. The function generates a multipanel graph with the spectrogram of the master sound file in top of that from test sound files, highlighting the position of correspondent test sounds on both in order to simplify assessing and adjusting their alignment:
[…]
