### 2024-01-12 49d29ba
create images on the fly vignette
