+++ tests/testthat/test-auto_realign.R
@@ -8,7 +8,7 @@ test_that("basic", {
   unaligned_test_sounds_est <- test_sounds_est
   set.seed(123)
   noise_time <- sample(c(0.009, -0.01, 0.03, -0.03, 0, 0.07, -0.007),
-  sum(unaligned_test_sounds_est$distance != 0),
+  nrow(unaligned_test_sounds_est),
   replace = TRUE)
   
   attr(unaligned_test_sounds_est, "check.res")$start <- 
