+++ vignettes/quantify_degradation.Rmd
@@ -39,14 +39,43 @@ editor_options:
 
 ```{r eval= TRUE, echo=FALSE}
 
-
 library(knitr)
-library(kableExtra)
-opts_chunk$set(
-  fig.align = "center",
-  message = FALSE,
-  warning = FALSE
-)
+
+# create custom printing method
+.print_df <- function(x, highlight = NULL, height = "300px", row.names = FALSE, ...) {
+  kbl <- kableExtra::kable(
+    as.data.frame(x)
+    ,
+    align = "c",
+    row.names = row.names,
+    format = "html",
+    escape = F
+  )
+  
+  if (!is.null(highlight))
+    kbl <-
+      kableExtra::column_spec(
+        kbl,
+        column = which(names(x) %in% highlight),
+        background = "#ccebff",
+        bold = TRUE
+      )
+  
+  kbl <-
+    kableExtra::kable_styling(kbl, bootstrap_options = "striped", font_size = 14)
+  
+  
+  kbl <-
+    kableExtra::scroll_box(kbl, width = "800px", height = height)
+  
+  asis_output(kbl)
+}
+
+
+opts_chunk$set(fig.align = "center",
+               message = FALSE,
+               warning = FALSE)
+
 ```
 
 <img src="baRulho_sticker.png" alt="baRulho logo" align="right" width = "25%" height="25%"/>
@@ -64,8 +93,8 @@ There are a few important things to keep in mind about functions for quantifying
 ## Required data structure
 
 The input data should contain some additional information. [baRulho](https://marce10.github.io/baRulho/) comes with an example `extended_selection_table` data set that can be used to understand the required data structure:
-s
-```{r,eval=TRUE}
+
+```{r,eval=FALSE}
 
 # load packages
 library(baRulho)
@@ -75,20 +104,24 @@ library(ggplot2)
 # load example data
 data("test_sounds_est")
 
+test_sounds_est
 ```
+
 
 
 ```{r, eval = TRUE, echo = FALSE}
 
-kbl <- kable(test_sounds_est, align = "c", row.names = F, format = "html", escape = F)
+# load packages
+library(baRulho)
+library(viridis)
+library(ggplot2)
 
-kbl <- column_spec(kbl, 7:9, background = "#ccebff", bold = TRUE)
+# load example data
+data("test_sounds_est")
 
-kbl <- kable_styling(kbl, bootstrap_options = "striped", font_size = 12)
 
-kbl <- scroll_box(kbl, height = "400px")
+.print_df(test_sounds_est, highlight = c("sound.id", "transect", "distance"))
 
-kbl
 ```
 &nbsp;
  
@@ -125,13 +158,10 @@ table(test_sounds_est$sound.id, test_sounds_est$distance)
 
 ```{r, eval = TRUE, echo=FALSE}
 
-tb <- table(test_sounds_est$sound.id, test_sounds_est$distance)
+tb <- as.data.frame.matrix(table(test_sounds_est$sound.id, test_sounds_est$distance))
 
-kbl <- kable(tb, align = "c", row.names = TRUE, format = "html", escape = F)
+.print_df(tb, height = NULL, row.names = TRUE)
 
-kbl <- kable_styling(kbl, bootstrap_options = "striped", font_size = 12)
-
-kbl
 ```
 
 
@@ -152,15 +182,8 @@ test_sounds_est
 
 ```{r, echo = FALSE}
 
[…]
