+++ vignettes/quantify_degradation.Rmd
@@ -187,7 +187,7 @@ test_sounds_est
 ```
 &nbsp;
 
-The function adds the column 'reference' which is then used by downstream functions plotting or measuring degradation. References are indicated as a the combination of the 'sound.files' and 'selec' column. For instance, '10m.wav-1' indicates that the row in which the 'selec' column is '1' and the sound file is '10m.wav' should be used as reference. The function also checks that the information 'X' is in the right format so it wont produce errors in downstream analysis (see 'X' argument description for details on format). The function will ignore rows with 'sound.id' 'ambient', 'start_marker' and 'end_marker'.
+The function adds the column 'reference' which is then used by downstream functions plotting or measuring degradation. References are indicated as a the combination of the 'sound.files' and 'selec' column. For instance, '10m.wav-1' indicates that the row in which the 'selec' column is '1' and the sound file is '10m.wav' should be used as reference. The function also checks that the information 'X' (the input annotation data) is in the right format so it wont produce errors in downstream analysis (see 'X' argument description for details on format). The function will ignore rows with 'sound.id' 'ambient', 'start_marker' and 'end_marker'.
 
 ### Visual inspection
 
