### 2024-01-12 95e4472
explain env.smooth
