### 2024-01-12 20a5fde
explain negative snr values
