### 2024-01-12 3828876
explain tsr negative values
