+++ man/excess_attenuation.Rd
@@ -36,7 +36,7 @@ consecutive windows, as in \code{\link[seewave]{spectro}}. Default is 50. Only u
 \item{path}{Character string containing the directory path where the sound files are found. Only needed when 'X' is not an extended selection table. If not supplied the current working directory is used.}
 }
 \value{
-Object 'X' with an additional column,  'excess.attenuation', containing the id of the sound used as reference and the computed excess attenuation values, respectively.
+Object 'X' with an additional column,  'excess.attenuation', containing the computed excess attenuation values (in dB).
 }
 \description{
 \code{excess_attenuation} measures excess attenuation in sounds referenced in an extended selection table.

[…]
