### 2024-01-12 #609 @maRce10
Many thanks for the review. I am really grateful for all your time and effort. I have made changes and responded to your comments/suggestions below, quoting the commit the change was made in.

## Response to @mikemahoney218

**comment:**  My main general comment about the package is, as Jeff and the automatic checks both identified, many functions in this package have a pretty high cyclomatic complexity. As an outside reviewer, this made it hard to trace any given function. I'd encourage the developers to break their functions down into more modular pieces, in order to possibly be able to reuse elements across functions but also to make the package easier to scan and maintain in the long run.

 _**response:**  I have broken down functions and now all functions have a cyclomatic compelxity lower than 15 maRce10/baRulho@b6868c4e41646927e5e81f3b4e7e291e24bc6ab7_

**comment:**  I saw the earlier conversation between you and Jeff on this, but I think the
size isn't due to using sound examples -- I think it's due to the size of images
you're using in your help files and vignettes. It seems like the PNGs in the
package are saved at 300 PPI resolution; saving them at 72 or 96 PPI instead
will save a ton of space (for instance, it shrinks the package sticker from 1.4
MB to 162.5 kB on my machine) while leaving them high-res enough for computer
displays. You might also consider moving the quantify_degradation vignette to be an
"article" instead, saving some of the space that gets used for those images.

 _**response:**  _ I have reduce the size of the sticker to 162.5 kB as suggested by the reviewer. (maRce10/baRulho@9c9d95dd425d50f1e0258b412f887779619657e5). I also remove 4 images from the vignettes as part of having all chunks evaluated (maRce10/baRulho@176266b43a73bcd01b16b59dc5c42abf26ee75ff & maRce10/ohun@49d29ba12f5f48d71683b5266633f54e8dc530f0). _

# Vignettes

## align_test_sounds.Rmd

**comment:**  In general, I'd encourage you to evaluate all chunks in a vignette unless
they're expected to error, on purpose, and this is highlighted to the reader.
I get into specifics below, but reading the rendered vignette I am confused by
what some functions are meant to output, and I believe there are some code
errors hidden in unevaluated chunks.

 _**response:** Now most vignette chuncks are evaluated, except for those that create image file which are shown in a gif or those with routines that take a considerable amount of time to run maRce10/baRulho@176266b43a73bcd01b16b59dc5c42abf26ee75ff & maRce10/baRulho@7fb6a5f7af1c7e9c456cc79e6268c76a787b632b_

**comment:**  This vignette defines td in a setup chunk that's not included in the vignette,
then uses it in user-visible code throughout the vignette. As a user, I'd expect
to be able to run the code displayed in the vignette. I'd suggest either
defining td in the user-visible part of the vignette, or using tempdir() as
an argument to these functions, rather than using the variable.

 _**response:**  _I have followed the recommendation by the reviewer to use tempdir() maRce10/baRulho@176266b43a73bcd01b16b59dc5c42abf26ee75ff_

**comment:**  The calls to spectrograms() and check_sels() using synth_master_annotations
error ("NAs found in bottom.freq and/or top.freq"). Those chunks are marked
as eval = FALSE, so maybe this is known, but there's nothing in the vignette
implying that these chunks will error -- in fact, there are spectrograms
included below the chunk, which mean I expect that running the code in that
chunk will produce the resulting graph.

 _**response:**  I have fixed these bugs. Now those code chunks are evaluated. I also decided to use a different spectrogram creating function (seewave::spectro()) that plots the spectrograms in the R graphic device, which is more intuitive maRce10/baRulho@176266b43a73bcd01b16b59dc5c42abf26ee75ff_

**comment:**  Is there any reason you switch from using arguments to set file paths to using
options right before find_markers()? The inconsistency confused me.

 _**response:**  options() is not longer used maRce10/baRulho@7fb6a5f7af1c7e9c456cc79e6268c76a787b632b_

**comment:**  You have a hidden call to print(TRUE) after calling
is_extended_selection_table(). I think it would probably be better to return
the actual output from is_extended_selection_table() instead. On my machine,
this returned FALSE, making the vignette incorrect (the actual class is
data.frame).

 _**response:**  _It has been removed and is_extended_selection_table() is evalauted maRce10/baRulho@7fb6a5f7af1c7e9c456cc79e6268c76a787b632b_

**comment:**  The command \code{\link{align_test_files}} in the manually fixing alignment
section produces a blank space in the HTML output. I think this should probably
be align_test_files().

 _**response:**  fixed maRce10/baRulho@58402dd843a35da85bf9779aa29652be378b1f09_

**comment:**  In auto_realign() I'm seeing:

Warning messages:
1: In attr(X, "check.results")$start[paste(X$sound.files, X$selec,  :
  number of items to replace is not a multiple of replacement length
2: In attr(X, "check.results")$end[paste(X$sound.files, X$selec, sep = "-") %in%  :
  number of items to replace is not a multiple of replacement length

Is that expected?

 _**response:**  Nope. I fixed it maRce10/baRulho@b47aef37d3f1b09f17f4003b1790c1ac765a42e7_

## quantify_degradation.Rmd

**comment:**  options(dest.path = td) (in the blur ratio section) errors because td isn't
defined. As with the other vignette, the mix of hidden evaluated code and
unevaluated displayed code means the actual code displayed in the vignette
doesn't always work.

 _**response:**  I have removed the use of options() and set all arguments on the function call maRce10/baRulho@598b18ed40db6f7ee3003da0ffa77e4470e3c9af_

**comment:**  There's a line in this vignette that says "The following code sets env.smooth = 800", but it's not obvious to me where or why that gets set.

 _**response:**  We have added env.smooth = 800 to the blur_ratio call and added that 'which produces smoother envelopes' maRce10/baRulho@3a8c41a691f6d41e23ccf788b8f9fe6bb48e3774_

**comment:**  Some of the plots use scale_color_viridis(), some use scale_color_manual(values = viridis::viridis()), and it's not obvious to me why this changes. It makes comparing the code in different chunks hard, as there are a number of things changing for reasons I can't identify.

 _**response:**  Fixed. Now all ggplot color scales use viridis directly maRce10/baRulho@7e22e55b795f3f58de150035735a7ab05a50bba2_

**comment:**  I don't know if it's your CSS formatting or something else, but while code in
the Rmd is formatted nicely, the ggplot2 code on the pkgdown site seems to all
be on a single line in a way that makes it hard to parse.

 _**response:**  I used opts_chunk$set(tidy = TRUE) but doesn't seem to work well. So I formatted the code myself (F2) maRce10/baRulho@7e22e55b795f3f58de150035735a7ab05a50bba2_

# Examples

**comment:**  If you're going to use \dontrun around examples that should work, which don't
highlight to users that they're expected to fail, I recommend using
devtools::run_examples(run_donttest = TRUE, run_dontrun = TRUE) to confirm
that the code actually will run successfully.

 _**response:**  I used devtools::run_examples(run_donttest = TRUE, run_dontrun = TRUE) to spot problematic examples and fix them accordingly maRce10/baRulho@d372f71da19f20c8dfd3f8bfa963c1aca5940e44. Note that manual_realign is an interacitve function_

**comment:**  blur_ratio.R: need to use ggplot2:: or library(ggplot2). It's probably
good practice to also wrap this in if (requireNamespace("ggplot2", quietly = TRUE)) (or, if you're willing to
pick up a dependency on rlang, rlang::check_installed("ggplot2")).

 _**response:** I added rlang::check_installed("ggplot2") and removed \dontrun maRce10/baRulho@2be41e78e675d428e543f43b8340a318b4ddf246_

**comment:** manual_realign.R: Error: object 'master_df' not found. Seems like Y should be as.data.frame(master_est) (or this object needs to be created earlier)? manual_realign.R: Error in eval(ei, envir) : object 'mako' not found. Need
to namespace viridis::mako().

 _**response:** master_df error fixed in maRce10/baRulho@d372f71da19f20c8dfd3f8bfa963c1aca5940e44. mako changed to viridis::mako() maRce10/baRulho@6459466444ae5431a13d1f7a69f045fce9b1503e_

 **comment:** spectrum_blur_ratio.R: also needs ggplot2:: or library(ggplot2).

 _**response:** I added rlang::check_installed("ggplot2") maRce10/baRulho@2be41e78e675d428e543f43b8340a318b4ddf246_

  **comment:** Warning messages:
1: In attenuation(f = 2000, dist = 50, dist0 = 1) :
  partial argument match of 'f' to 'frequency'
2: In match.call(definition, call, expand.dots, envir) :
  partial argument match of 'f' to 'frequency'
3: In attr(X, "check.results")$start[paste(X$sound.files, X$selec,  :
  number of items to replace is not a multiple of replacement length
4: In attr(X, "check.results")$end[paste(X$sound.files, X$selec, sep = "-") %in%  :
  number of items to replace is not a multiple of replacement length
5: In eg$dur : partial match of 'dur' to 'duration'
6: In eg$harm : partial match of 'harm' to 'harmonics'
7: In eg$harm : partial match of 'harm' to 'harmonics'

 _**response:** maRce10/baRulho@e1fdd34736a525d48ef6222bab13f0ee1b9fd8cc_

 **comment:** blur_ratio.R: Any reason the call to blur_ratio() on line 42 is commented out?

 _**response:** Not really. I uncommented it maRce10/baRulho@e1916dad1da71d8d405704d84290ce2e5b87b50f_

  **comment:** Is there a reason you're using the @usage tag? Normally this section is
auto-generated by rOxygen2, ensuring it stays up-to-date and properly formatted.

 _**response:**  Did not know rOxygen2 can do it. I removed all function @usage fields maRce10/baRulho@9582da689ccd92e9861c81827de1d5f3b658f625_

 **comment:** I love the standardization of arguments across functions; I think that's a clear
strength of the package. You might consider using rOxygen2 templates (see the
relevant section of the packaging guide) to centralize these arguments' documentation, in order to reduce the amount of
places you need to touch to update any documentation.

 _**response:** I always wanted to do this but was not sure how. I added argument templates to a new R code file template_params.R which is now referenced to by other functions using @inheritParams template_params maRce10/baRulho@0852fc0594de8cc0ceda438ae75021aafb2e3d91_

  **comment:** It might be nice to define more of the constants used in attenuation(), so
that the calculations are a bit easier to interpret. Right now, there's a lot
of magic numbers in this function call.

 _**response:** The complexity of the calculations makes difficult to explain those numbers so I pointed out to the original sources and added comments to calculations done in the function.
maRce10/baRulho@3610d04e995051a0baeb324420a91d7b2d52ceb2_

 **comment:** In attenuation(), I think dist0 and dist are required inputs, but are
placed after optional arguments. It might be nicer to place these arguments
after frequency, to allow users to pass the distance arguments via position.

 _**response:** I placed dist0 and dist after frequency maRce10/baRulho@2cad8a200bafa1a816ea78f780d57ff638276034_

  **comment:** I think this is probably the easiest time to get rid of deprecated arguments
and functions, before the package gets to CRAN/gets more visibility via the
rOpenSci organization. I see the parallel argument and output argument in
several functions is deprecated, and I think this is a good time to drop
those, as well as the search_templates() and spcc_distortion() functions.

 _**response:** I get rid of all deprecated arguments and functions maRce10/baRulho@26a66796a75471e69d4d05c4383affffab2f42ed_

 **comment:** auto_realign(): I'd recommend providing the list of options to wn here.
See the Enumerate possible options in the "Tidy design principles" book for an example of why this is useful.

 _**response:** Done maRce10/baRulho@8b4ce4e733ad67aad8a83310547395bf4b931c0f_

  **comment:** I'd similarly suggest using enumerations for noise.ref, type, method,
and other functions where users must provide an option from a pre-defined
list. Then use arg.match() or rlang::match_arg() to ensure provided values
match one of the options.

 _**response:** Done maRce10/baRulho@d382a7f3e1d1b359442137b9c56102c90d147b55_

 **comment:** blur_ratio(), plot_aligned_sounds(): I think it'd make sense to mention in the help file that wl must be even, and that it will be adjusted if not. Also, does wl need to be an integer value? It's documented as a numeric vector, but any decimal values will have 1 added to them.

 _**response:** Now it states that "@param wl a vector with a single even integer number specifying the window length of the spectrogram, default is \code{NULL}. If supplied, 'hop.size' is ignored. Odd integers will be rounded up to the nearest even number." maRce10/baRulho@c08e626b6438a5ec931e40276af54f13b6e8aec5_

  **comment:** set_reference_sounds(): I see that this function requires eight columns with specific names. Is there any reason to not let the user control, via function arguments, which columns are used?

 _**response:** The columns are required for the function to set which sounds will be used as reference. Most of those columns are also used by other functions, so it is very convenient to standardize the names of the columns. _

 **comment:** Consider having functions that generate plots (I'm thinking of plot_degradation(), plot_blur_ratio()) return the file paths they generated. I was surprised a few times when a function appeared to not return anything, because it had written a figure out to disk instead.

 _**response:** I added a message to indicate where the files where saved and the functions also return the file paths  invisibly (without printing). maRce10/baRulho@3bf53d80869f1f3111dbf859b9cf6629c0ce5312_

  **comment:** plot_blur_ratios(): colors says it should be a Character vector of length 4,
but then the default argument is viridis::viridis(3), of length 3. Not sure
what's correct here.

 _**response:** Fixed: "@param colors Character vector of length 4 containing the colors to be used for the color to identify the reference sound  (element 1),  the color to identify the test sound (element 2) and the color of blurred region (element 3)" maRce10/baRulho@c18f56a1382088d4cfeaa4dd25752168910b7075_

 **comment:** tail_to_signal_ratio(): bp having a default value that doesn't have the
same type or length as what users would provide is a bit confusing. Maybe this
default should be NULL?.

 _**response:** I decided to keep it as it is because the function let users to change the default value ("frange") which uses the actual frequency range in the input data to specific ranges_

  **comment:** type seems to be used in different ways in several different functions, with different sets of options in each. Consider using different argument names for each.

 _**response:** I renames type to snr.formula and tsr.formula signal_to_noise_ratio and tail_to_signal_ratio respectively maRce10/baRulho@9c3980709ffe56e8c8ce364bba516f76e728812b_

 **comment:** The contributing guide seems to be missing an email:

    You must never report security related issues, vulnerabilities or bugs including sensitive information to the issue tracker, or elsewhere in public. Instead sensitive bugs must be sent by email to <>. I'd either provide an email or drop this part of the guide.

 _**response:** Done maRce10/baRulho@ffe7432898165c513bbc85c2cbe04b3cb0bccd58_

 ## Response to @DenaJGibbon

  **comment:** For this round of comments I focused mostly on the vignette. I was able to work through all of the examples without any issue. However, there is one component that is missing from this. I think it will be important to add a section or separate page that focuses on how to run this using your own dataset. For example, if folks start with Raven selection tables, sounds files, and metadata of distance and other information, how do they get started using these methods? The examples in the vignette work great, but I think some work could be done to help people get their own data in the pipeline with this.

 _**response:** The "aligned_test_sounds.Rmd" vignette uses actual recordings as those that people will get in their own experiments:

 "We will use acoustic data included in the package as example data. This is a data subset from a sound transmission experiment. The complete data set of the experiment is hosted in this repository"

 "Their names are self-explanatory: a master sound file (“master.wav”) and its annotations, a reference sound file (“reference_1m.wav”) and two test files (“test_10m.wav” and “test_20m.wav”) re-recorded at 10 and 20 m respectively.

Now we will save those files in a temporary working directory (those objects are extended selection tables in the R environment) to resemble the usual case in which the acoustic data is found in sound files."

 I have added that: "Note the 'sound.id' column which contains a unique identifier for each sound. This will be used to label their counterparts in the test sound files. Users can import annotations from [Raven](https://www.ravensoundsoftware.com/software/raven-pro/) using the [Rraven](https://cran.r-project.org/package=Rraven) package ([see this vignette](https://CRAN.R-project.org/package=Rraven/vignettes/Rraven.html))."  maRce10/baRulho@f17d56aa07235325083afe2221be928ba4d47460_

 **comment:** It would be helpful to include an example or type of research question. This may be one of the first exposure people have to this type of work, so a few examples would be helpful

 _**response:** I added that "These experiments aim to answer research questions such as: How habitat structure has shaped the transmission properties of animal acoustic signals? Which acoustic features are shaped by selection for improving transmission? What is the active space of a acoustic signal?"  maRce10/baRulho@6d8cc6d51124115e05fb94d835cfa6b5ff8b8c15_

 **comment:** What does the check part mean? A figure legend describing the diagram would be helpful.

 _**response:** I slightly modified the diagram, adding the related functions to it. So I ended up removing the "check" part. I also added the following legend:
 "Figure 1. Diagram depicting a typical workflow of experiments working on signal tranmission and degradation. Nodes with black font indicate steps that can be conducted using baRulho functions. Blue nodes denote the functions that can be used at those steps." maRce10/baRulho@95ba615f22873c12ad914cbb475b687b1fbb9086_

  **comment:**  I was quite impressed with how smoothly I was able to work through this section of the vignette- nice work!

   _**response:** thanks!_

  **comment:**  You might want to include a warning about playing the sound synthetic master file (!). It gave me quite a shock!

 _**response:** I added: "(be careful when playing the master sound file as it can be shocking for some audiences!)" maRce10/baRulho@78c0526d1699d7a36c53db15eb2fb2334530d522_

  **comment:** I can imagine there will be times when spectrogram cross-correlation won't work to time align files. Perhaps adding something about using caution and the need to manually check would be good.

 _**response:** I have clarified that "In some cases alignments might not be accurate, particularly when markers have been considerably degraded on the test sound files" maRce10/baRulho@_

 **comment:** I realize this is an R package vignette and not a peer-reviewed paper, but I think including some pros/cons of different measures would be helpful. How should someone who is coming to this for the first time decide which values to measure and report?

 _**response:** I am working on the manuscript at this point, which touches upon this issues. I will refer to the paper once it is publish._

[…]
