### 2024-01-22 #609 @mikemahoney218
Looking at the current `main` version, I'm having a number of issues related to `tail_to_signal_ratio()`. `devtools::check()` fails with:
```
   Quitting from lines 509-513 [unnamed-chunk-38] (quantify_degradation.Rmd)
   Error: processing vignette 'quantify_degradation.Rmd' failed with diagnostics:
   unused argument (tsr.formula = 1)
   --- failed re-building ‘quantify_degradation.Rmd’
```

Deleting that section of the vignette, I then get:
```
❯ checking examples ... [21s/21s] ERROR
  Running examples in ‘baRulho-Ex.R’ failed
  The error most likely occurred in:

  > base::assign(".ptime", proc.time(), pos = "CheckExEnv")
  > ### Name: tail_to_signal_ratio
  > ### Title: Measure reverberations as tail-to-signal ratio
  > ### Aliases: tail_to_signal_ratio
  >
  > ### ** Examples
  >
  > {
  +   # load example data
  +
  +   data("test_sounds_est")
  +
  +   # set global options
  +   options(pb = FALSE)
  +
  +   # using margin for noise of 0.01
  +   tsr <- tail_to_signal_ratio(X = test_sounds_est, mar = 0.01)
  +
  +   # use tsr.formula 2 which is equivalent to tail-to-noise ratio
  +   tsr <- tail_to_signal_ratio(X = test_sounds_est, mar = 0.01, tsr.formula = 2)
  + }
  Error in check_arguments(fun = arguments[[1]], args = arguments) :
    could not find function "check_arguments"
  Calls: tail_to_signal_ratio
  Execution halted

❯ checking for code/documentation mismatches ... WARNING
  Codoc mismatches from documentation object 'tail_to_signal_ratio':
  tail_to_signal_ratio
    Code: function(X, mar, parallel = NULL, cores = getOption("mc.cores",
                   1), pb = getOption("pb", TRUE), type = 1, bp =
                   "freq.range", output = NULL, hop.size =
                   getOption("hop.size", 1), wl = getOption("wl", NULL),
                   ovlp = getOption("ovlp", 0), path =
                   getOption("sound.files.path", "."))
    Docs: function(X, mar, cores = getOption("mc.cores", 1), pb =
                   getOption("pb", TRUE), tsr.formula = 1, bp =
                   "freq.range", hop.size = getOption("hop.size", 1), wl
                   = getOption("wl", NULL), ovlp = getOption("ovlp", 0),
                   path = getOption("sound.files.path", "."))
    Argument names in code not in docs:
      parallel type output
    Argument names in docs not in code:
      tsr.formula
    Mismatches in argument names (first 3):
      Position: 3 Code: parallel Docs: cores
      Position: 4 Code: cores Docs: pb
      Position: 5 Code: pb Docs: tsr.formula

❯ checking installed package size ... NOTE
    installed size is  8.5Mb
    sub-directories of 1Mb or more:
      doc    5.7Mb
      help   1.6Mb

❯ checking R code for possible problems ... [10s/10s] NOTE
  tail_to_signal_ratio: no visible global function definition for
    ‘check_arguments’
  tail_to_signal_ratio: no visible global function definition for
    ‘report_assertions2’
    check_arguments report_assertions2
```

The changes seem promising so far, but is it possible to fix these errors so that I can do a more complete re-review?
