### 2024-01-22 #609 @mikemahoney218
Is your pkgdown website not rendering anymore? The version on the website is out-of-date from your actual Rmd files, and contains a few errors that aren't present in the Rmd.

In the alignment vignette, `warbleR::is_extended_selection_table(aligned_tests)` still returns `FALSE` for me (when running only user-visible code), while a hidden chunk prints `TRUE`. Not sure what's going on here -- the fact that this apparently _should_ be an extended selection table, but isn't, _feels_ concerning? But I have no idea if this is actually a problem.

I'm still seeing one error in examples:
```
# add_noise.R
> data("test_sounds_est")

> # make it a 'by element' extended selection table
> X <- warbleR::by_element_est(X = test_sounds_est)
  |                                                  | 0 % ~calculating  Error in wave[a:b, ] : only 0's may be mixed with negative subscripts
```

Those :point_up: are the only blocking issues I've got left. I'd also consider using `checkmate::assert_integerish()` inside `.check_wl()` to confirm `wl` is integer-ish (that is, "is a whole number (even if it's stored as a numeric)"), but I don't think it's absolutely necessary. Thanks for the thorough revision!
