### 2024-01-22 1c79390
fix class of output aligned_test
