### 2024-01-23 4630f55
added assert_integerish() to .check_wl()
