### 2024-01-23 d53d24a
fix check_intigerish
