+++ DESCRIPTION
@@ -7,7 +7,7 @@ Maintainer: Marcelo Araya-Salas <marcelo.araya@ucr.ac.cr>
 Description: Intended to facilitate acoustic analysis of (animal) sound transmission experiments, which typically aim to quantify changes in signal structure when transmitted in a given habitat by broadcasting and re-recording animal sounds at increasing distances. The package offers a workflow with functions to prepare the data set for analysis as well as to calculate and visualize several degradation metrics, including blur ratio, signal-to-noise ratio, excess attenuation and envelope correlation among others (Dabelsteen et al 1993 <doi:10.1121/1.406682>).
 License: GPL (>= 2)
 Imports: utils, stats, seewave, tuneR, fftw, methods, viridis, Sim.DiffProc, png, checkmate, cli, rlang
-Depends: R (>= 3.2.1), warbleR (>= 1.1.29), ohun (>= 1.0.1)
+Depends: R (>= 3.2.1), warbleR (>= 1.1.30), ohun (>= 1.0.1)
 LazyData: TRUE
 URL: https://github.com/maRce10/baRulho
 BugReports: https://github.com/maRce10/baRulho/issues
