### 2024-01-24 20fd8c5
Deploy from Github Actions build 7635436420 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/7635436420
Commit: e7b6720b57e781ae0777a0868175350e2a2305ba
