### 2024-01-24 #609 @maRce10
**comment:** Is your pkgdown website not rendering anymore? The version on the website is out-of-date from your actual Rmd files, and contains a few errors that aren't present in the Rmd.

 _**response:**  Not sure why github pages is not building the site. Any help with that is welcome_

**comment:** In the alignment vignette, warbleR::is_extended_selection_table(aligned_tests) still returns FALSE for me (when running only user-visible code), while a hidden chunk prints TRUE. Not sure what's going on here -- the fact that this apparently should be an extended selection table, but isn't, feels concerning? But I have no idea if this is actually a problem.

 _**response:**  I fixed the code. The object class is now evaluated. The class is data frame maRce10/baRulho@1c79390a0b49f01ed8c781d69ac6f989e17b9442_

**comment:**I'm still seeing one error in examples:

# add_noise.R
> data("test_sounds_est")

> # make it a 'by element' extended selection table
> X <- warbleR::by_element_est(X = test_sounds_est)
  |                                                  | 0 % ~calculating  Error in wave[a:b, ] : only 0's may be mixed with negative subscripts

 _**response:**  This was a problem with a dependency, warbleR. Fortunately, I am the maintainer of that package so I updated it on CRAN and udpated the required package version in DESCRIPTION.  maRce10/baRulho@d53d24a808f47d901840a13af1aacbd5a2f0c468_

**comment:** I'd also consider using checkmate::assert_integerish() inside .check_wl() to confirm wl is integer-ish (that is, "is a whole number (even if it's stored as a numeric)"), but I don't think it's absolutely necessary.

 _**response:** Done maRce10/baRulho@4630f5556668d5a9b7692ede30318a1141f54a87_

thanks!
