### 2024-01-24 #609 @jhollist
@maRce10 Thanks for submitting your response to @mikemahoney218 feedback.

As a side note, I think the review bot commands need to be submitted in a comment without anything else in that comment.  Try that again and see if it works.

@DenaJGibbon could you take a look at the response to your review in https://github.com/ropensci/software-review/issues/609#issuecomment-1889567333 and let us know if there any outstanding concerns?
