### 2024-01-24 #609 @mikemahoney218
Opened https://github.com/maRce10/baRulho/pull/18 to hopefully fix the website, but otherwise all my concerns have been addressed and I'm comfortable recommending acceptance :smile:
