### 2024-01-24 d954c74
Deploy from Github Actions build 7642747401 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/7642747401
Commit: afe34d5b6b55d997670c0e60ae7d914072d2fd84
