### 2024-01-24 #609 @maRce10
> Opened [maRce10/baRulho#18](https://github.com/maRce10/baRulho/pull/18) to hopefully fix the website, but otherwise all my concerns have been addressed and I'm comfortable recommending acceptance 😄

yes it fixed the website. Thanks a lot!
