### 2024-01-27 c2196fd
Deploy from Github Actions build 7676281757 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/7676281757
Commit: afe34d5b6b55d997670c0e60ae7d914072d2fd84
