### 2024-01-29 e321ae3
Deploy from Github Actions build 7691085735 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/7691085735
Commit: afe34d5b6b55d997670c0e60ae7d914072d2fd84
