### 2024-02-02 356c853
Deploy from Github Actions build 7751312625 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/7751312625
Commit: afe34d5b6b55d997670c0e60ae7d914072d2fd84
