### 2024-02-04 d1a96d0
Deploy from Github Actions build 7771576894 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/7771576894
Commit: afe34d5b6b55d997670c0e60ae7d914072d2fd84
