### 2024-02-05 550612c
Deploy from Github Actions build 7779493641 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/7779493641
Commit: afe34d5b6b55d997670c0e60ae7d914072d2fd84
