### 2024-02-08 6a1f3a2
Deploy from Github Actions build 7824792319 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/7824792319
Commit: afe34d5b6b55d997670c0e60ae7d914072d2fd84
