### 2024-02-08 #609 @DenaJGibbon
Hello @maRce10 .

I have had an enjoyable afternoon working through "baRulho" using my own data. I am at the 'excess attenuation' function, and in the vignette it says: "Two methods for calculating excess attenuation are provided (see ‘method’ argument)." However, the function returns an error when I try to specify a method and there is nothing about the method in the function documentation.  Could you please clarify? Does this mean based on the two reference methods, or are there two ways to calculate EA with the package?

Below is the output from the 'plot_degradation' function on my own data (gibbons from Borneo). It looks great!

![plot_degradation_p1](https://github.com/ropensci/software-review/assets/26265144/ab3417ea-2840-431c-9644-8da182b6b550)

I am still working through all the functions and will probably not be able to submit the full review until next week. Thanks for your patience!
