### 2024-02-08 69fafb8
update vignette removing mention of additional EA methods
