### 2024-02-08 #609 @maRce10
Thanks @DenaJGibbon !

There was an argument in excess_attenuation() to select the type of EA to calculate. However I removed that option as one method was not very reliable. However I forgot to update the vignette.  I Just did that in this commit: maRce10/baRulho@69fafb847ae4a012889b0d1f6a117a279651ca4d
