### 2024-02-09 d0fcd6c
Deploy from Github Actions build 7839538625 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/7839538625
Commit: 69fafb847ae4a012889b0d1f6a117a279651ca4d
