### 2024-02-12 17e9126
Deploy from Github Actions build 7867203361 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/7867203361
Commit: 69fafb847ae4a012889b0d1f6a117a279651ca4d
