### 2024-02-15 88a248e
Deploy from Github Actions build 7911009295 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/7911009295
Commit: 69fafb847ae4a012889b0d1f6a117a279651ca4d
