### 2024-02-16 72a4087
Deploy from Github Actions build 7925935047 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/7925935047
Commit: 69fafb847ae4a012889b0d1f6a117a279651ca4d
