### 2024-02-20 fcb1586
Deploy from Github Actions build 7968381360 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/7968381360
Commit: 69fafb847ae4a012889b0d1f6a117a279651ca4d
