### 2024-02-21 f0347e2
Deploy from Github Actions build 7983902285 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/7983902285
Commit: 69fafb847ae4a012889b0d1f6a117a279651ca4d
