### 2024-02-22 #609 @DenaJGibbon
First, I want to say you have done a great job getting all these functions together in a relatively easy to digest format. This will be a highly useful tool for folks working in bioacoustics. I worked through all of the example using my own data, and have some specific comments based on that experience below. I also apologize for the delay - I have been playing catch up since I got back from the field. I should be able to respond sooner in the future.

# Getting set up
I spent much of the time I allotted to do this review getting my data in the correct format. As I suggested in my previous review, I really think that to lower the barrier to entry for folks you should include step-by-step instructions on how to get the data into the correct format. I referred to the 'warbleR' documentation on the extended selection tables, but it was not clear (from reading the documentation) how to get the information about distance, sound.id, etc. into the table. After lots of trial and error I was able to get it working, but for folks that don't have a ton of coding experience this will be hard for them.

# Plot degradation
It would be great if you could include bit more information in the description for envelope and spectrum - I dug into the code to see how they were calculated but it should be easier to see. Also if there is some way to include the axis labels or units of the two in the plot (or maybe default printed somewhere) that would be good.

# Excess attenuation
I am a bit confused about the example in the EA section, as a proportion of the values are negative (which would indicate amplification of the sound). Could you please double check and let me know what is going on here? I also think that including the equation in the documentation would be helpful here.

# Noise profiles
I think there is some missing info: "The function uses internally to calculate frequency spectra. "
