### 2024-02-24 b856812
Deploy from Github Actions build 8028150726 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/8028150726
Commit: 69fafb847ae4a012889b0d1f6a117a279651ca4d
