### 2024-02-25 6b9543b
Deploy from Github Actions build 8035443345 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/8035443345
Commit: 69fafb847ae4a012889b0d1f6a117a279651ca4d
