### 2024-02-27 d348a90
Deploy from Github Actions build 8059529601 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/8059529601
Commit: 69fafb847ae4a012889b0d1f6a117a279651ca4d
