### 2024-03-02 93c0658
fix missing link to seewave::meanspec
