+++ R/tail_to_signal_ratio.R
@@ -21,7 +21,7 @@
 #' with the tail-to-signal ratio values (in dB).
 #' @export
 #' @name tail_to_signal_ratio
-#' @details Tail-to-signal ratio (TSR) measures the ratio of power in the tail of reverberations to that in the test sound. A general margin in which reverberation tail will be measured must be specified. The function will measure TSR within the supplied frequency range (e.g. bandpass) of the reference sound ('bottom.freq' and 'top.freq' columns in 'X'). Two methods for computing reverberations are provided (see 'tsr.formula' argument). Note that 'tsr.formula' 2 is not equivalent to the original description of TSR in Dabelsteen et al. (1993) and  is better referred to as tail-to-noise ratio. Tail-to-signal ratio values are typically negative as signals tend to have higher power than that in the reverberating tail.
+#' @details Tail-to-signal ratio (TSR) measures the ratio of power in the tail of reverberations to that in the test sound. A general margin in which reverberation tail will be measured must be specified. The function will measure TSR within the supplied frequency range (e.g. bandpass) of the reference sound ('bottom.freq' and 'top.freq' columns in 'X'). Two methods for computing reverberations are provided (see 'tsr.formula' argument). Note that 'tsr.formula' 2 is not equivalent to the original description of TSR in Dabelsteen et al. (1993) and  is better referred to as tail-to-noise ratio. Tail-to-signal ratio values are typically negative as signals tend to have higher power than that in the reverberating tail. TSR can be ~0 when both tail and signal have very low amplitude.
 #' @examples {
 #'   # load example data
 #'
@@ -195,9 +195,16 @@ tail_to_signal_ratio <- function(X,
         # sound (or noise) RMS
         sig_RMS <- seewave::rms(sig.env)
         
-        # get reference ambient noise RMS
+        # convert to 0.0001 if sig_RMS is 0 to avoid errors in STR measurements
+        if (sig_RMS == 0)
+          sig_RMS <- 0.0001
         
+        # get reference ambient noise RMS
         tail_RMS <- seewave::rms(tail.env)
+        
+        # convert to 0.0001 if tail_RMS is 0 to avoid errors in STR measurements
+        if (tail_RMS == 0)
+          tail_RMS <- 0.0001
         
         # Calculate tail.to.signal ratio
         str <- tail_RMS / sig_RMS
