+++ vignettes/align_test_sounds.Rmd
@@ -68,8 +68,9 @@ opts_chunk$set(fig.align = "center",
 
 <img src="baRulho_sticker.png" alt="baRulho sticker" align="right" width = "25%" height="25%"/>
 
-
 This vignette shows deals with two things: how to create synthetic sounds for playback experiments and how to align re-recorded sounds to their reference once playback has been conducted. Here is some terms that will be used in this vignettes and throughout the package.
+
+&nbsp;
 
 <div class="alert alert-info">
 

[…]
