### 2024-03-02 262ad55
Deploy from Github Actions build 8120295475 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/8120295475
Commit: 93c0658e2593007cc892e31f97574f0d97ec79f1
