### 2024-03-04 e6e4412
explain envelope/spectrum axis
