### 2024-03-04 35e4406
add formula description to EA
