### 2024-03-04 #609 @maRce10
Thanks a lot @DenaJGibbon for your very useful comments! Here are the responses:

**comment:** I spent much of the time I allotted to do this review getting my data in the correct format. As I suggested in my previous review, I really think that to lower the barrier to entry for folks you should include step-by-step instructions on how to get the data into the correct format. I referred to the 'warbleR' documentation on the extended selection tables, but it was not clear (from reading the documentation) how to get the information about distance, sound.id, etc. into the table. After lots of trial and error I was able to get it working, but for folks that don't have a ton of coding experience this will be hard for them.

 _**response:**  You are totally right, this is an issue I have had in mine for a while, mostly related to the package warbleR which is where these different input data formats are defined. baRulho basically take all 3 data formats used by the package warbleR, which I also maintain. So I added a new vignette to warbleR, called "annotation data format" (https://marce10.github.io/warbleR/articles/annotation_data_format.html), to have a more detailed explanation of the structure of these objects.  I have added a link to this vignette to  the  section "Require data structure" in the  vignette "Quantify degradation" in baRulho. I have also added definitions for all the required columns in the input data description in "Quantify degradation". maRce10/baRulho@626e999ffa951e69e31fe1b32cfc013f06590273_

**comment:** Excess attenuation

I am a bit confused about the example in the EA section, as a proportion of the values are negative (which would indicate amplification of the sound). Could you please double check and let me know what is going on here? I also think that including the equation in the documentation would be helpful here.

 _**response:**  I have revised the code (thanks to  input from users) and the EA calculation is now more accurate. Now  only 2 measures are negative in the example maRce10/baRulho@176266b43a73bcd01b16b59dc5c42abf26ee75ff.
I have also added that  " Excess attenuation is computed as \code{-20 * log10(rms("test signal") / rms("reference signal"))) - (20 * log10(1 / "distance")} in which 'rms(..)' represents the root mean square of an amplitude envelope." to the function description maRce10/baRulho@35e4406caf0d0e21b75ee7cbd8620c3bc64f4745_

**comment:** Plot degradation

It would be great if you could include bit more information in the description for envelope and spectrum  I dug into the code to see how they were calculated but it should be easier to see. Also if there is some way to include the axis labels or units of the two in the plot (or maybe default printed somewhere) that would be good.

_**response:**  I have added to the function description that "Amplitude envelopes and power spectra are computed using the functions  \code{\link[seewave]{envelope}} (warbleR package) and \code{\link[seewave]{spec}} (seewave package) respectively. This two visualizations show the power distribution in time and frequency between the minimum and maximum power values for each sound. Therefore scales are not necessarily comparable across panels." maRce10/baRulho@e6e441237ce6a04d747a9fdb8c691a56d082d22c. I think adding scales would be tricky as plots are already pretty busy. I also think this scales are not extremely informative if sound files are not calibrated as in most cases. __

**comment:**  Noise profiles

I think there is some missing info: "The function uses internally to calculate frequency spectra. "

 _**response:**  fixed!  maRce10/baRulho@e6e441237ce6a04d747a9fdb8c691a56d082d22c_

thanks!
