### 2024-03-05 cc241fa
fix broken links to figures
