+++ vignettes/align_test_sounds.Rmd
@@ -68,7 +68,7 @@ opts_chunk$set(fig.align = "center",
 
 <img src="baRulho_sticker.png" alt="baRulho sticker" align="right" width = "25%" height="25%"/>
 
-This vignette shows deals with two things: how to create synthetic sounds for playback experiments and how to align re-recorded sounds to their reference once playback has been conducted. Here is some terms that will be used in this vignettes and throughout the package.
+This vignette deals with two things: how to create synthetic sounds for playback experiments and how to align re-recorded sounds to their reference once playback has been conducted. Here is some terms that will be used in this vignettes and throughout the package.
 
 &nbsp;
 
@@ -470,13 +470,25 @@ aligned_imgs
 
 ```
 
- These are two of the files generated by the code above:
+```{r, echo = FALSE}
+
+# try to copy files to man/figures
+# fc <- file.copy(from = aligned_imgs[1:2], 
+#           to = file.path("../man/figures", basename(aligned_imgs[c(1, 4)])))
+
+fc <- file.copy(from = aligned_imgs[c(1, 4)], 
+          to = file.path("../vignettes", basename(aligned_imgs[c(1, 4)])),overwrite = TRUE)
+
+```
+
+
+These are two of the files generated by the code above:
 
 ```{r, echo=FALSE, out.width="100%"}
 
-knitr::include_graphics(aligned_imgs[[1]])
+knitr::include_graphics(basename(aligned_imgs[1]))
 
-knitr::include_graphics(aligned_imgs[[4]])
+knitr::include_graphics(basename(aligned_imgs[4]))
 
 ```
 
