### 2024-03-05 0a1fb85
improve colors diagram
