### 2024-03-05 9485f26
clarified package assumptions for measuring degradation
