+++ vignettes/quantify_degradation.Rmd
@@ -78,21 +78,28 @@ opts_chunk$set(fig.align = "center",
 
 ```
 
+<!-- light brown box -->
+<div class="alert alert-warning">
+
 <img src="baRulho_sticker.png" alt="baRulho logo" align="right" width = "25%" height="25%"/>
 
-
-There are a few important things to keep in mind about functions for quantifying degradation:
+These are key considerations regarding the package behavior when assessing degradation:
 
 - The package currently assumes that all recordings have been made with the same equipment and recording volume. This will be modified in future versions to allow for amplitude calibration of recordings.
 - Wave envelope and frequency spectrum calculations are made after applying a bandpass filter within the frequency range of the reference sound ('bottom.freq' and 'top.freq' columns)
 - The package offers two methods to compare sounds to the reference: 
     1. Compare all sounds with the counterpart that was recorded at the closest distance to source (e.g. compare a sound recorded at 5m, 10m and 15m with its counterpart recorded at 1m). This is the default method.
     1. Compare all sounds with the counterpart recorded at the distance immediately before (e.g. a sound recorded at 10m compared with the one recorded at 5m, then sound recorded at 15m compared with the one recorded at 10m and so on).
+- The function will try to use references from the same transect. However, if there is another test sound from the same 'sound.id' at a shorter distance in other transects, it will be used as reference instead. This behavior aims to account for the fact that in this type of experiments reference sounds are typically recorded at 1 m and at single transect. 
 - The package assumes all sound files have the same sampling rate.
+</div>
+
+
 
 ## Required data structure
 
-The input data should contain some additional information. [baRulho](https://marce10.github.io/baRulho/) comes with an example annotation data set that can be used to understand the required data structure:
+ 
+Transmission experiments tend to follow a common experimental design in which model sounds are re-recorded at increasing distance within a transect. Hence, the data must indicate, besides the basic acoustic annotation information (e.g. sound file, time, frequency), the transect and distance within that transect for each sound. [baRulho](https://marce10.github.io/baRulho/) comes with an example annotation data set that can shows the required data structure:
 
 ```{r,eval=FALSE}
 
@@ -106,8 +113,6 @@ data("test_sounds_est")
 
 test_sounds_est
 ```
-
-
 
 ```{r, eval = TRUE, echo = FALSE}
 
@@ -124,7 +129,7 @@ data("test_sounds_est")
 
 ```
  
-Transmission experiments tend to follow a common experimental design in which model sounds are re-recorded at increasing distance within a transect. Hence, the data must indicate, besides the basic acoustic annotation information (e.g. sound file, time, frequency), the transect and distance within that transect for each sound. These are required input data columns:
+This the despcription of the required input data columns:
 
  1) **sound.files**: character or factor column with the name of the sound files including the file extension (e.g. "rec_1.wav")
  1) **selec**: numeric, character or factor column with a unique identifier (at least within each sound file) for each annotation (e.g. 1, 2, 3 or "a", "b", "c")
@@ -140,7 +145,9 @@ Transmission experiments tend to follow a common experimental design in which mo
 
 # Setting reference sounds 
 
-The combined information from these columns is used to identify the reference sounds for each test sound. The function `set_reference_sounds()` does exactly that. There are two possible experimental designs when defining reference sounds (which is controlled by the argument 'method' in `set_reference_sounds()`):
+The combined information from these columns is used to identify the reference sounds for each test sound. The function `set_reference_sounds()` does exactly that. Hence, unless you define the reference sound for each test sound manually, `set_reference_sounds()` must always be run before any degradation measuring function. 
+
[…]
