### 2024-03-05 ecf283b
improve reference setting description
