### 2024-03-06 3409bdc
minor details vignettes
