### 2024-03-06 e8cf4b8
Update quantify_degradation.Rmd
