### 2024-03-06 7a556da
Update quantify_degradation.Rmd
