+++ vignettes/align_test_sounds.Rmd
@@ -73,9 +73,10 @@ This vignette deals with two things: how to create synthetic sounds for playback
 
 <div class="alert alert-info">
 
-<img src="baRulho_sticker.png" alt="baRulho sticker" align="right" width = "25%" height="25%"/>
 
 # Glossary 
+
+<img src="baRulho_sticker.png" alt="baRulho sticker" align="right" width = "25%" height="25%"/>
 
 -**Model sound**: sound in which transmission properties will be studied, usually found in the original field recordings or synthetic sound files.
 
