### 2024-03-06 e37774d
Deploy from Github Actions build 8166374406 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/8166374406
Commit: 540e32f6c08920ea6543fd66a058745b4998d887
