### 2024-03-06 bb7eb99
Deploy from Github Actions build 8166509428 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/8166509428
Commit: 2fb8e9ae7df7892ecd272cf51f97d6cb832e6726
