### 2024-03-12 e3b7c92
Deploy from Github Actions build 8243492370 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/8243492370
Commit: 2fb8e9ae7df7892ecd272cf51f97d6cb832e6726
