### 2024-03-13 54e5a02
Deploy from Github Actions build 8259377465 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/8259377465
Commit: 2fb8e9ae7df7892ecd272cf51f97d6cb832e6726
