### 2024-03-14 #609 @jhollist
Thanks for the heads up @maRce10.  Not sure why these keep slipping by!

@DenaJGibbon everything look good to you in https://github.com/ropensci/software-review/issues/609#issuecomment-1977695348
