### 2024-03-15 06c4686
Deploy from Github Actions build 8291215240 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/8291215240
Commit: 2fb8e9ae7df7892ecd272cf51f97d6cb832e6726
