### 2024-03-16 d86a69b
Deploy from Github Actions build 8305254272 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/8305254272
Commit: 2fb8e9ae7df7892ecd272cf51f97d6cb832e6726
