### 2024-03-17 741dab9
Deploy from Github Actions build 8312878615 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/8312878615
Commit: 2fb8e9ae7df7892ecd272cf51f97d6cb832e6726
