### 2024-03-18 6ebd424
Deploy from Github Actions build 8321180460 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/8321180460
Commit: 2fb8e9ae7df7892ecd272cf51f97d6cb832e6726
