### 2024-03-19 6ef42ff
Deploy from Github Actions build 8337454240 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/8337454240
Commit: 2fb8e9ae7df7892ecd272cf51f97d6cb832e6726
