### 2024-03-21 f1b0858
Deploy from Github Actions build 8369594120 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/8369594120
Commit: 2fb8e9ae7df7892ecd272cf51f97d6cb832e6726
