### 2024-03-24 dd7d694
Deploy from Github Actions build 8406798043 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/8406798043
Commit: 2fb8e9ae7df7892ecd272cf51f97d6cb832e6726
