### 2024-03-25 #609 @DenaJGibbon
A few more comments - the package is in pretty good shape! Nice job!

## Vignette
- The printout of class(est) appears to have distances in it, but there is not an explanation in the text about where they came from in this vignette (https://marce10.github.io/warbleR/articles/annotation_data_format.html)
- I find the printout sections in the vignette page hard to follow/understand. Perhaps another call like head or colnames would provide the info needed and be cleaner?

## Package
- In the 'add_noise' function description it would be helpful to describe what kind of noise (e.g, white noise?)
- The 'detection_distance' function is in the package, but I couldn't find anything in the vignette about it (unless am I missing something?). This is a cool feature, but I couldn't find much info on it. And when I run the example the output is not intuitive.
- I ran through all the (~22) function examples, and they all worked - nice job!
