### 2024-03-28 6cbda41
Deploy from Github Actions build 8462322378 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/8462322378
Commit: 2fb8e9ae7df7892ecd272cf51f97d6cb832e6726
