+++ NAMESPACE
@@ -61,5 +61,7 @@ importFrom(stats,approx)
 importFrom(stats,ave)
 importFrom(stats,complete.cases)
 importFrom(stats,cor)
+importFrom(stats,fft)
 importFrom(stats,na.omit)
+importFrom(stats,rnorm)
 importFrom(stats,runif)
