### 2024-03-28 e6b07b1
update description of detection distance
