### 2024-03-28 #609 @maRce10
Thanks again @DenaJGibbon for your useful comments! Here are the responses:

**comment:** The printout of class(est) appears to have distances in it, but there is not an explanation in the text about where they came from in this vignette (https://marce10.github.io/warbleR/articles/annotation_data_format.html)

_**response:**  There was an issue with the ANSI format of the selection table printing method. But it is already fixed  maRce10/warbleR@3f3efc3934ddd1902fece4efdca8cfadd69fef9d_

**comment:** I find the printout sections in the vignette page hard to follow/understand. Perhaps another call like head or colnames would provide the info needed and be cleaner?

 _**response:**  I think this was also fixed with the changes made when dealing with the previous comment_

**comment:** In the 'add_noise' function description it would be helpful to describe what kind of noise (e.g, white noise?)

_**response:**   It adds white noise. But I just added a new argument "kind" that can be used for using different (white, pink, brown, red, and power).  maRce10/baRulho@e41319953a24aa3e404fbe83b3c47b1497a981b6_

**comment:**  The 'detection_distance' function is in the package, but I couldn't find anything in the vignette about it (unless am I missing something?). This is a cool feature, but I couldn't find much info on it. And when I run the example the output is not intuitive.

 _**response:**  I added this to the function description: "The function computes the maximum distance at which a sound would be detected, which is calculated as the distance in which the sound pressure level (SPL) goes below the specified SPL cutoff ('spl.cutoff')). This is returned as an additional column 'detection.distance' (in m). " maRce10/baRulho@e6b07b137ae3c56913357b71ed31181e6878f9e9_

**comment:**  I ran through all the (~22) function examples, and they all worked - nice job!

 _**response:** sweeet!

thanks!

(@jhollist)
