### 2024-03-28 bf92fb4
Deploy from Github Actions build 8474294542 [ci skip]
Build URL: https://github.com/maRce10/baRulho/actions/runs/8474294542
Commit: e6b07b137ae3c56913357b71ed31181e6878f9e9
